Finite State Machine

We offer here an easy way to construct a Finite State Machine (FSM).

Finite State Machine is a common structure used in a hardware design. Offering a simple method to construct a FSM may be useful.

Procedure of designing a FSM

Construct a FSM

First construct a FSM defining name of states and the machine itself.

VerilogWriter.FSM — Method

FSM(name, state::String...)

Create an FSM object with name name, whose states are state....

VerilogWriter.FSM — Method

FSM(name, states::Vector{String})

Create an FSM object with name name, whose states are states.

Add transition rules

Add a new rule with transadd!.

VerilogWriter.transadd! — Method

transadd!(x::FSM, cond::Wireexpr, newtrans::Pair{String, String})

Add a new transition rule for the state machine x.

The new rule here is:

The transition from state newtrans[1] to state newtrans[2]

This transition occures when cond is true and the current state is newtrans[1]

Examples

julia> fsm = FSM("nstate", "uno", "dos", "tres"); # create a FSM

julia> transadd!(fsm, (@wireexpr b1 == b2), "uno" => "dos"); # transition from "uno" to "dos" when "b1 == b2"

julia> vshow(fsmconv(Case, fsm));
case (nstate)
    uno: begin
        if ((b1 == b2)) begin
            nstate <= dos;
        end
    end
    dos: begin
        
    end
    tres: begin
        
    end
endcase
type: Case

As shown above cond is the condition to be true when the transition occurs, and as newtrans argument you assign a Pair object of strings, "src. state" => "dest. state".

Convert from FSM to verilog codes

We offer methods to generate several components in verilog, which compose FSM structure.

They are

A case statement for transition between states

reg declaration that holds the current state information.

localparams each of which represents one state in the FSM.

VerilogWriter.fsmconv — Method

fsmconv(::Type{Onedecl}, x::FSM)

Generate one reg declaration in verilog that holds the state value at the time.

VerilogWriter.fsmconv — Method

fsmconv(::Type{Localparams}, x::FSM)

Generate localparams that declares the value which corresponds to each state.

VerilogWriter.fsmconv — Method

fsmconv(::Type{Case}, x::FSM)

Convert FSM object into Case object.

You may also prewiew all of these with vshow(fsm).

struct FSM description

FSM(name::String, states::Vector{String}, transcond::Array{Wireexpr}, transvalid::Array{Bool})

VerilogWriter.FSM — Type

Struct for a Finite State Machine, not the Flying Spaghetti Monster.

Each column of matrices corresponds to the source state, and row for the destination state. Not supposed to directly edit the content of matrices without using handlers.

[
    false false 
    true  false
]

in transvalid means transition from the 1st state to the 2nd state is provided.