Basic Automation

We offer some tools to automatically add additional information inferred from a given Verilog-like codes.

Reset in Always Blocks

Given a content of always blocks, you may automatically reset all wires which appear at the LHS in the block.

VerilogWriter.autoreset — Function

autoreset(x::Ifcontent; clk=defclk, rst=defrst, edge=posedge, reg2d::Dict{String, Wireexpr}=Dict{String, Wireexpr}())

Given x::Ifcontent, returns always_ff/always block that resets every wire/regs appear at Lhs of x.

This is synchronous reset. reg2d is a pair of "name of wire" and "ram depth in wireexpr".

autoreset(x::Alwayscontent; clk=defclk, rst=defrst, edge=posedge, reg2d::Dict{String, Wireexpr}=Dict{String, Wireexpr}())

Automatically reset wires which appear in x::Alwayscontent.

Sensitivity list in the original Alwayscontent will be ignored.

Example

c = @always (
    r1 <= r2;
    if b1 
        r2 <= 0
        r3 <= r3 + r4
    else 
        r3 <= 0
    end
) 
r = autoreset(c; clk=(@wireexpr clk), rst=(@wireexpr ~resetn))
vshow(r)

# output

always_ff @( posedge clk ) begin
    if ((~resetn)) begin
        r1 <= 0;
        r2 <= 0;
        r3 <= 0;
    end else begin
        r1 <= r2;
        if (b1) begin
            r2 <= 0;
            r3 <= (r3 + r4);
        end else begin
            r3 <= 0;
        end
    end
end
type: Alwayscontent

autoreset(x::Vmodule; clk=defclk, rst=defrst, reg2d::Dict{String, Wireexpr}=Dict{String, Wireexpr}())

Return a new Vmodule object whose Alwayscontents are all reset.

Automatic Wire Declaration

VerilogWriter.autodecl — Function

autodecl(x, env::Vmodenv)::Vmodenv

Declare wires in x which are not yet declared in env. Raise error when not enough information to determine width of all wires is given.

Type of x is restricted to the type wireextract! accepts.

Examples

Inference Success

pts = @ports (
        @in 16 din;
        @in b1
)
env = Vmodenv(pts)

c = @ifcontent (
    reg1 = 0;
    reg2 = din;
    if b1 
        reg1 = din[10:7]
    end
) 

venv = autodecl(c, env)
vshow(venv)

# output

input [15:0] din
input b1

logic [3:0] reg1;
logic [15:0] reg2;
type: Vmodenv

You may also declare ports/wires beforehand whose width is unknown.

When declaring ports/wires without specifying its bit width, assign -1 as its width.

ps = @ports (
    @in 2 x;
    @in -1 y;
    @out @reg A z
) 
ds = @decls (
    @wire -1 w1;
    @wire B w2
)

ab = @always (
    z <= r1 + r2 + r3;
    r4 <= (y & w1) << r1[1:0];
    r5 <= y + w2
)
env = Vmodenv(Parameters(), ps, Localparams(), ds)
nenv = autodecl(ab.content, env)

vshow(nenv)

# output

input [1:0] x
input [B-1:0] y
output reg [A-1:0] z

wire [B-1:0] w1;
wire [B-1:0] w2;
logic [A-1:0] r1;
logic [A-1:0] r2;
logic [A-1:0] r3;
logic [B-1:0] r4;
logic [B-1:0] r5;
type: Vmodenv

Fail in Inference

c = @ifcontent (
    reg1 = 0;
    reg2 = din;
    if b1 
        reg1 = din[10:7]
    end
) 

autodecl(c)

# output

ERROR: Wire width cannot be inferred for the following wires.
1. b1
2. reg2 = din

autodecl(x)::Vmodenv

Conduct wire width inference under an empty environment.

autodecl(x::Vmodule)::Vmodule

Using ports, parameters, localparams, decls in x::Vmodule as an environment, conduct wire width inference and return a new Vmodule object with inferred wires.

env in an argument for autodecl is of type Vmodenv.

VerilogWriter.vfinalize — Function

vfinalize(x::Vmodule)

Return new Vmodule object generated by applying autoreset and autodecl on x.

reg2d is automatically generated before calling autodecl. This is based on the observation that depth of RAM can not be inferred usually.