Reference

All docstrings are listed below.

VerilogWriter.eachfieldconstruct — Method

eachfieldconstruct(strc)

Make contructors for struct strc.

Example

For

struct S 
a::A
b::B
...
end

constructors

S(x::A) = S(x, B(),...)
S(x::B) = S(A(), x,...)

will be generated. Note that A(), B() should return appropriate objects.

VerilogWriter.explicitCallop — Constant

Must be explicitly called as function e.g. |(wire1), &(wire2).

VerilogWriter.noJuliaop — Constant

&& and || are invalid identifier in julia.

VerilogWriter.wirenoname — Constant

name == "" for no name (intermediate node)

VerilogWriter.wirevalinvalid — Constant

-1 for Int field of Wireexpr which should not contain any data.

VerilogWriter.Alassign — Type

Assign statement inside always blocks.

VerilogWriter.Alwayscontent — Type

Represent always blocks.

VerilogWriter.Assign — Type

Assign one statement.

VerilogWriter.Atype — Type

Type of always blocks.

VerilogWriter.Case — Type

case (cnd)
    cnd1: begin
        foo1
    end
    cnd2: begin
        foo2
    end
    ...
endcase

converts to

Case(cnd, [
    (cnd1 => foo1),
    (cnd2 => foo2),
    ...
])

VerilogWriter.Decls — Type

Multiple wire declarations.

VerilogWriter.Edge — Type

Edge in sensitivity lists.

VerilogWriter.Ifcontent_inner — Type

Container of one if-block (, one elseif block, or one else block).

Parametrized by T, U only for mutual recursion with Ifelseblock and Case, thus used as Ifcontent_inner{Ifelseblock, Case}, which is aliased as Ifcontent.

VerilogWriter.Ifelseblock_inner — Type

Container of an if-else block. Parametrized by T for mutual recursion. Used as Ifelseblock = Ifelseblock_inner{Case}.

VerilogWriter.Localparams — Type

Multiple localparams.

VerilogWriter.Onedecl — Type

Represent one wire declaration.

VerilogWriter.Onelocalparam — Type

One localparam.

VerilogWriter.Oneparam — Type

Parameter in verilog.

VerilogWriter.Oneport — Type

Represent a single port declaration.

VerilogWriter.Parameters — Type

List of parameters.

VerilogWriter.Portdirec — Type

Port direction object.

VerilogWriter.Ports — Type

Gather multiple ports.

VerilogWriter.Sensitivity — Type

Sensitivity list for always blocks.

VerilogWriter.Vmodinst — Type

Vmodinst

Represents module instantiation.

i = Vmodinst(
    "mod1",
    "u1",
    [
        "A" => Wireexpr(15)
    ],
    [
        "p1" => (@wireexpr x << 1),
        "p2" => (@wireexpr y)
    ]
)
vshow(i)

# output

mod1 #(
    .A(15)
) u1 (
    .p1((x << 1)),
    .p2(y)
);
type: Vmodinst

VerilogWriter.Vmodule — Type

Represents one verilog module.

VerilogWriter.Wireexpr — Type

Wire expressions in verilog.

Contains unnecessary information to handle all the wires in the same type.

One motivation to avoid making different types of wires (e.g. using parametric types) is that it seemed beneficial to have multiple wires in a vector, and that it is sometime a disadvantage in performance to use abstract types when creating vectors.

Some operators on wires in Verilog, which are listed in Enum Wireop, are overloaded for Wireexpr. Note that reduction operators (unary &, |, ^) are not in Julia, and logical and, or (&&, ||) can be applied only for booleans in Julia, thus are not available as an operator for Wireexpr objects.

Examples

julia> w1 = @wireexpr x + y;

julia> w2 = @wireexpr z;

julia> vshow(w1 & w2);
((x + y) & z)
type: Wireexpr

julia> w3 = @wireexpr x;

julia> w4 = @wireexpr y && z; # && is available inside `@wireexpr` and `wireexpr` methods.

julia> vshow(w3 | w4)
(x | (y && z))
type: Wireexpr

VerilogWriter.Wireop — Type

Verilog operators.

Unary &, | does not exist so explicitly call as function e.g. &(wire), |(wire) (& behaves in a wickedmanner...?), and are only available inside quoted expressions. Xor in verilog ^ is in Julia exponential operator, and the difference in an association exists.

VerilogWriter.Wiretype — Type

Wiretype object.

Base.convert — Method

Convert parameters and localparams with each other.

Base.convert — Method

Convert oneparam and onelocalparam with each other.

Base.convert — Method

Convert oneparam and onelocalparam with each other.

Base.convert — Method

Convert parameters and localparams with each other.

Base.hash — Method

Base.hash(x::Wireexpr, h::UInt)

Hash for Wireexpr to make it acceptable as keys for Dict.

Base.isequal — Method

Base.:isequal(uno::Wireexpr, dos::Wireexpr)

Equality for Wireexpr.

Note that Base.:(==) for Wireexpr is defined to create a new Wireexpr object.

Base.iterate — Method

Base.iterate(x::Wireexpr)

Needed to deal with some return value that may be either Wireexpr or Vector{Wireexpr}. May cause type instability, but it is already inevitable in parsing AST.

VerilogWriter.TAB — Constant

TAB as 4 spaces.

VerilogWriter.indent — Method

Indent txt with spaces of length $eachspalen*depth$.

VerilogWriter.newlineconcat — Method

Concatenate uno and dos with '\n' inserted inbetween.

VerilogWriter.showfield — Method

for documentation of structs.

VerilogWriter.space — Method

generate spaces of length x.

VerilogWriter.spacewrap — Method

Wrap txt with space of length width.

VerilogWriter.vshow — Method

vshow(x; systemverilog=true)

Print the structs in VerilogWriter.jl in a readable format (except for enums). When systemverilog=false, output always, always @* instead of always_ff, always_comb, respectively.

VerilogWriter.addatype! — Method

addatype!(x::Alwayscontent)

Infer type of always block (combinational or sequential) and add the information to x.

VerilogWriter.always — Method

always(expr::Expr)

Parse AST into always block as Alwayscontent using ralways.

Also infers type of always using addatype!.

Syntax

<oneblock>;<oneblock>[;<oneblock>;...]

<oneblock> is the expression that can be parsed by oneblock. ; in between <oneblock>s are strictly needed.

@posedge <wirename>; <ifelsestatements>/<assignments>

Set sensitivity list using macro syntax. @negedge is also possible. You must put @posegde/@negedge statement at the beginning, and only once.

Examples

a1 = always(:(
    w1 = w2;
    if b2 
        w1 = w3 
    end
))
vshow(a1)

# output

always_comb begin
    w1 = w2;
    if (b2) begin
        w1 = w3;
    end
end
type: Alwayscontent

a1 = always(:(
    @posedge clk;
    
    if b1 == b2
        w1 <= w2 + w3 
    else
        w1 <= ~w1 
    end
))
vshow(a1)

# output

always_ff @( posedge clk ) begin
    if ((b1 == b2)) begin
        w1 <= (w2 + w3);
    end else begin
        w1 <= ~w1;
    end
end
type: Alwayscontent

VerilogWriter.declmerge — Method

declmerge(d::Decls...)

Merge multiple Decls objects into one Decls.

VerilogWriter.declmerge — Method

declmerge(d::Localparams...)

Merge multiple Localparams objects into one Localparams.

VerilogWriter.declmerge — Method

declmerge(d::Parameters...)

Merge multiple Parameters objects into one Parameters.

VerilogWriter.declmerge — Method

declmerge(d::Ports...)

Merge multiple Ports objects into one Ports.

VerilogWriter.declmergegen — Method

declmergegen()

Declare declmerge for multiple types.

VerilogWriter.ifadd! — Method

ifadd!(ifblock::Ifelseblock, cond, ifcont)

Add cond and ifcont at the head of ifblock.

Intended to add if-clause to Ifelseblock constructed from elseif-clause (because elseif-clause often parses first).

VerilogWriter.invport — Method

invport(onep::Oneport)

Return Oneport object whose directions are reversed from onep. Wiretype information (reg, wire, logic) is lost when inverted for reg cannot be at input port.

VerilogWriter.invports — Method

invports(ps::Ports)

Return Ports object whose directions are reversed from ps.

Examples

pts = @ports (
    @in 8 bus1, bus2;
    @out @reg bus3
)
ipts = invports(pts)
vshow(pts)
vshow(ipts)

# output

(
    input [7:0] bus1,
    input [7:0] bus2,
    output reg bus3
);
type: Ports
(
    output [7:0] bus1,
    output [7:0] bus2,
    input bus3
);
type: Ports

VerilogWriter.@sym2wire — Macro

@sym2wire(arg::Expr)

Declare new Wireexprs of name arg..., respectively.

Example

julia> @sym2wire x, y, z;

julia> vshow(y);
y
type: Wireexpr

VerilogWriter.@sym2wire — Macro

@sym2wire(arg::Symbol)

Declare new Wireexpr of name arg.

VerilogWriter.Wireexpr — Method

Wireexpr(n::Int)

Literal of Integer.

VerilogWriter.Wireexpr — Method

Wireexpr(w::Integer, n::Integer)

Literal with width specification, printed in the decimal format.

VerilogWriter.Wireexpr — Method

Wireexpr(n::String)

Case where no slice is needed.

VerilogWriter.Wireexpr — Method

Wireexpr(n::Symbol)

Convert Symbol to String and construct Wireexpr.

VerilogWriter.Wireexpr — Method

Wireexpr(expr::Wireexpr)

Return the argument itself, used in slice construction to make it possible to apply the same method to msb as Int and as Wireexpr.

VerilogWriter.Wireexpr — Method

Wireexpr(op::Wireop, w::Wireexpr...)

Apply an operation of w wires.

Examples

julia> a = @wireexpr a; b = @wireexpr b;

julia> c = Wireexpr(add, a, b); # equivalent to wireexpr(:($(a) + $(b)))

julia> vshow(c);
(a + b)
type: Wireexpr

julia> d = Wireexpr(redor, a); vshow(d); # wireexpr(:(|($(a))))
|(a)
type: Wireexpr

VerilogWriter.Wireexpr — Method

Wireexpr(op::Wireop, v::Vector{Wireexpr})

Apply an operation on wires in v.

VerilogWriter.Wireexpr — Method

Wireexpr()

Create an empty expression.

VerilogWriter.Wireexpr — Method

Wireexpr(n::String, msb::T1, lsb::T2) where {T1 <: Union{Int, Wireexpr}, T2 <: Union{Int, Wireexpr}}

Slice of n from msb to lsb.

Index for wires can be Int or Wireexpr.

VerilogWriter.Wireexpr — Method

Wireexpr(n::String, msb::T) where {T <: Union{Int, Wireexpr}}

Get one bit at msb from n.

VerilogWriter.blockconvable — Type

Types to which AST of block can be converted.

VerilogWriter.blockconv — Method

blockconv(vas, vif, vwire, ::Val{Ifcontent})

Convert lists of Alassign, Ifelseblock, and Wireexpr to a single Ifcontent. Wireexpr is supposed to be empty.

VerilogWriter.blockconv — Method

blockconv(vas, vif, vwire, ::Val{Ifelseblock})

Convert lists of Alassign, Ifelseblock, and Wireexpr to a single Ifelseblock (whose conds is empty). Wireexpr is supposed to be empty.

VerilogWriter.blockconv — Method

blockconv(vas, vif, vwire, ::Val{Wireexpr})

Convert input to a single Wireexpr.

It is expected that both vas::Vector{Alassign} and vif::Vector{Ifelseblock} are empty, and vwire is of length 1. Only applied to block at elseif.args[1] (condition of elseif-clause).

VerilogWriter.decloneline — Method

decloneline(expr::Expr)

Parse Julia AST into wire declaration as Vector{Onedecl}.

The number of Onedecl objects returned may differ according to the number of wires declared in one line (e.g. input dout <=> input din1, din2, din3).

Syntax

Similar to that of portoneline.

@wire/@reg/@logic [<width>] <wirename1>[, <wirename2>,...]

Examples

julia> d = decloneline(:(@reg 10 d1)); vshow(d);
reg [9:0] d1;
type: Onedecl

julia> d = decloneline(:(@logic 8 d1,d2,d3)); vshow(d);
logic [7:0] d1;
type: Onedecl
logic [7:0] d2;
type: Onedecl
logic [7:0] d3;
type: Onedecl

julia> d = decloneline(:(@wire A >> 2 w1, w2)); vshow(d);
wire [(A >> 2)-1:0] w1;
type: Onedecl
wire [(A >> 2)-1:0] w2;
type: Onedecl

VerilogWriter.decloneline — Method

decloneline(expr::Ref{T}) where {T}

For macro call.

VerilogWriter.decloneline_inner — Method

decloneline_inner(wt, wid::Wireexpr, vars::Expr)

Declaration of multiple wires (e.g. logic x, y, z).

VerilogWriter.decloneline_inner — Method

decloneline_inner(wt, wid::Wireexpr, var::Symbol)

Declaration of a single wire (e.g. wire [1:0] d1, reg clk).

VerilogWriter.decls — Method

decls(expr::Expr, ::Val{:block})

Multiple lines of wire declarations.

VerilogWriter.decls — Method

decls(expr::Expr, ::Val{:macrocall})

Parse single line of declaration.

VerilogWriter.decls — Method

decls(expr::Expr)

Parse Julia AST into wire declaration section object Decls.

Syntax

<onedecl>[;<onedecl>;...]

Multiple decloneline expressions which are concatenated by ; can be accepted.

Example

d = decls(:(
    @wire w1;
    @reg 8 w2,w3,w4;
    @logic 32 w5
))
vshow(d)

# output

wire w1;
reg [7:0] w2;
reg [7:0] w3;
reg [7:0] w4;
logic [31:0] w5;
type: Decls

VerilogWriter.decls — Method

decls(expr::Decls...)

For interpolation of Decls objects.

VerilogWriter.decls — Method

decls(expr::Vector{Onedecl}...)

For interpolation of a single vector.

VerilogWriter.decls — Method

decls(expr::Vector{Onedecl})

For insertion with metaprogramming.

Example

a = @decloneline @reg 8 x1, x2
b = decls(:(
    $(a);
    @wire y1, y2
))
vshow(b)

# output

reg [7:0] x1;
reg [7:0] x2;
wire y1;
wire y2;
type: Decls

VerilogWriter.decls — Method

decls(expr::Ref{T}) where {T}

For macro calls.

VerilogWriter.ifcontent — Method

ifcontent(x::Expr)

Convert into Ifcontent what is convertible to Alwayscontent.

Example

x = @ifcontent (
    a = b;
    if b 
        x = c
    else
        x = d 
    end
)
vshow(x)

# output

a = b;
if (b) begin
    x = c;
end else begin
    x = d;
end
type: Ifcontent

VerilogWriter.ifcontent — Method

ifcontent(x::Ref{T}) where {T}

For macro call.

VerilogWriter.localparams — Method

localparams(expr::Expr)

Multiple localparams.

Examples

julia> p = localparams(:(x = 10)); vshow(p);
localparam x = 10;
type: Localparams

p = localparams(:(
    a = 111;
    b = 222;
    c = 333
))
vshow(p)

# output

localparam a = 111;
localparam b = 222;
localparam c = 333;
type: Localparams

VerilogWriter.localparams — Method

localparams(expr::Ref{T}) where {T}

For macro call.

VerilogWriter.oneblock — Method

oneblock(expr, ::Val{:(=)})

Assignment in Julia, stands for combinational assignment in Verilog.

VerilogWriter.oneblock — Method

oneblock(expr, ::Val{:<=})

Parse expr whose head is call and expr.args[1] is <=.

Parse as a comparison opertor in Julia, interpret as sequential assignment in Verilog.

VerilogWriter.oneblock — Method

oneblock(expr, ::Val{:call})

Dispatch methods of oneblock according to expr.args[1].

VerilogWriter.oneblock — Method

oneblock(expr, ::Val{:elseif})

Parse expr into Ifelseblock where expr is elseif-clause.

VerilogWriter.oneblock — Method

oneblock(expr::Expr)

Parse AST in Julia grammar to one if-else statement or one assignment inside always blocks in Verilog, which are Ifelseblock and Alassign, respectively.

As using Julia expression and Julia parser as if it is of Verilog, there can be difference in grammatical matters between what can be given as the expr and real Verilog (e.g. operator precedence of <=).

Syntax

<wirename1> = <wireoperation>

One blocking assignment. (Note that bit width cannot be specified yet, nor can it be printed.) <wireoperation> is a expression accepted by wireexpr.

If-else statement

if <wireoperation>
    <oneassignment>
    <oneassignment>
    ...
elseif <wireoperation> 
    <oneassignment>
    <ifelsestatement>
    ...
else
    <ifelsestatement>
    ...
end

If-else statement written in 'Julia syntax', not in Verilog syntax, can be accepted. else block and elseif are not compulsory. Since if end are at the top level no ; inside if-else statement is needed. Nested if-else statement can be also accepted as in usual Julia.

Examples

julia> a1 = oneblock(:(w1 <= w2)); 

julia> vshow(a1);
w1 <= w2;
type: Alassign

julia> a2 = oneblock(:(w3 = w4 + ~w2)); vshow(a2);
w3 = (w4 + ~w2);
type: Alassign

a3 = oneblock(:(
    if b1 == b2
        w5 = ~w6
        w7 = w8 
    elseif b2 
        w9 = w9 + w10
    else
        if b3 
            w11 = w12 
        end
    end
))
vshow(a3)

# output

if ((b1 == b2)) begin
    w5 = ~w6;
    w7 = w8;
end else if (b2) begin
    w9 = (w9 + w10);
end else begin
    if (b3) begin
        w11 = w12;
    end
end
type: Ifelseblock

VerilogWriter.oneblock — Method

oneblock(expr::T) where {T <: Union{Alassign, Ifelseblock}}

For insertion through metaprogramming.

Example

a = @oneblock r = s & t
b = @oneblock (
    if b 
        x = y 
    else
        x = z
    end
)
c = always(:(
    p = q;
    $(a);
    $(b)
))
vshow(c)

# output

always_comb begin
    p = q;
    r = (s & t);
    if (b) begin
        x = y;
    end else begin
        x = z;
    end
end
type: Alwayscontent

VerilogWriter.oneblock — Method

oneblock(expr::T) where {T <: Union{Symbol, Int}}

Convert expr to Wireexpr. Needed in parsing block.

VerilogWriter.oneblock — Method

oneblock(expr::UInt8)

UInt8 may be given when user writes e.g. 0b10, 0x1f. Used when parsing block

VerilogWriter.oneblock — Method

oneblock(expr, ::Val{:block}, ::Val{T}) where {T <: blockconvable}

Parse expr (which is block) into T. block appears at various nodes, thus need to explicitly indicate which type the return value should be. Types that is allowed as return value is in blockconvable.

See also blockconv.

VerilogWriter.oneblock — Method

oneblock(expr, ::Val{:elseif}, ::Val{T}) where {T <: blockconvable}

When given Val{T} as the third argument, if expr is not block (= if the second argument is not Val{:block}), ignore Val{T}.

This is needed in oneblock(expr, ::Val{:elseif}), for expr, whose expr.head == :elseif, can have either block or elseif as its expr.args[3]. Checks that the return value is of type T.

VerilogWriter.oneblock — Method

oneblock(expr, ::Val{T}) where {T <: blockconvable}

Helper method to make it possible to dispatch oneblock without giving expr.head as an argument.

VerilogWriter.oneblock — Method

oneblock(expr::Ref{T}) where {T}

Dereference and parse expr given by user. Helper function for macro @oneblock.

See also @oneblock.

VerilogWriter.onelocalparam — Method

onelocalparam(expr::Expr)

One localparam object. The syntax is the same as oneparam.

Example

julia> vshow(onelocalparam(:(x = 100)))
localparam x = 100;
type: Onelocalparam

VerilogWriter.onelocalparam — Method

onelocalparam(expr::Ref{T}) where {T}

For macro call.

VerilogWriter.oneparam — Method

oneparam(expr::Expr)

One parameter.

Syntax

<paramname> = <val::Int>

julia> vshow(oneparam(:(x = 10)));
parameter x = 10
type: Oneparam

VerilogWriter.oneparam — Method

oneparam(expr::Ref{T}) where {T}

For macro call.

VerilogWriter.parameters — Method

parameters(expr::Expr)

Multiple parameters.

Using convert and localparams inside.

Example

v = parameters(:(
    x = 10;
    y = 20;
    z = 30
))
vshow(v)

# output

#(
    parameter x = 10,
    parameter y = 20,
    parameter z = 30
)
type: Parameters

VerilogWriter.parameters — Method

parameters(expr::Ref{T}) where {T}

For macro call.

VerilogWriter.pdirecsym — Method

pdirecsym(sym)

Convert symbols @in, @out to pin, pout (::Portdirec).

VerilogWriter.portoneline — Method

portoneline(expr::Expr)

Parse Julia AST to one line of port declaration as Oneport.

Syntax

@in <wirename>, @out <wirename>

One port declaration of width 1.

@in/@out <wirename1, wirename2, ...>

Multiple port declaration of width 1 in one line.

@in/@out <width> <wirename1[, wirename2, ...]>

Port declarations of width <width>.

@in/@out @<wiretype> [<width>] <wirename1[, wirename2, ...]>

Port declaration with wiretypes [of width <width>].

Examples

julia> p1 = portoneline(:(@in din));

julia> vshow(p1);
input din
type: Oneport

julia> p2 = portoneline(:(@out din1, din2, din3)); vshow(p2);
output din1
type: Oneport
output din2
type: Oneport
output din3
type: Oneport

julia> p3 = portoneline(:(@in 8 din)); vshow(p3);
input [7:0] din
type: Oneport

julia> p4 = portoneline(:(@out @reg 8 dout)); vshow(p4);
output reg [7:0] dout
type: Oneport

julia> p5 = portoneline(:(@out (A+B)<<2 x, y)); vshow(p5); # width with parameter
output [((A + B) << 2)-1:0] x
type: Oneport
output [((A + B) << 2)-1:0] y
type: Oneport

VerilogWriter.portoneline — Method

portoneline(expr::Ref{T}) where {T}

For macro call.

VerilogWriter.ports — Method

ports(expr::Expr, ::Val{:block})

Contain multiple lines of port declaration.

VerilogWriter.ports — Method

ports(expr::Expr, ::Val{:macrocall})

Contain only single line declaration.

VerilogWriter.ports — Method

ports(expr::Expr)

Convert Julia AST into port declarations as Ports object.

Syntax

<portoneline>[;<portoneline>;...]

Multiple lines of portoneline expressions separated by ; can be accepted.

Example

pp = ports(:(
    @in p1;
    @in @wire 8 p2, p3, p4;
    @out @reg 2 p5, p6
))

vshow(pp)

# output

(
    input p1,
    input [7:0] p2,
    input [7:0] p3,
    input [7:0] p4,
    output reg [1:0] p5,
    output reg [1:0] p6
);
type: Ports

VerilogWriter.ports — Method

ports(expr::Ports...)

Interpolation of one Vector{Ports}, nothing else given. e.g. ports(:([ports(:(#= none:1 =# @in(6, x))), ports(:(#= none:1 =# @in(y)))]...))

VerilogWriter.ports — Method

ports(expr::Vector{Oneport})

Insertion through metaprogramming.

Example

a = @portoneline @in 8 d1, d2, d3
b = ports(:(
    @in d0;
    $(a);
    @out @reg 8 dout
))
vshow(b)

# output

(
    input d0,
    input [7:0] d1,
    input [7:0] d2,
    input [7:0] d3,
    output reg [7:0] dout
);
type: Ports

VerilogWriter.ports — Method

ports(expr::Ref{T}) where {T}

For macro call.

VerilogWriter.ralways — Method

ralways(expr::Expr, ::Val{:block})

Convert multiple ifelse-blocks and assigns to one always-block.

VerilogWriter.ralways — Method

ralways(expr::Expr)

Convert Julia AST to Verilog always-block.

Does not infer type of alwaysblock here. Dispatches methods here according to expr.head.

VerilogWriter.ralways — Method

ralways(expr::Expr, ::T) where {T <: Val}

The case where expr.head is not block, which means expr is one assign (inside always-block) or if-else block.

VerilogWriter.ralways — Method

ralways(expr::Ref{T}) where {T}

For macro call.

VerilogWriter.ralwayswithSensitivity — Method

ralwayswithSensitivity(expr)

Parse always block with sensitivity list (e.g. @posedge clk).

VerilogWriter.removelinenumbernode — Method

removelinenumbernode(args::Vector{T}) where {T}

Remove LineNumberNode from args.

VerilogWriter.wireexpr — Method

wireexpr(expr, ::Val{:(:)}, ::Val{:call})

Parse wire slice with range object, e.g. x[a:b].

VerilogWriter.wireexpr — Method

wireexpr(expr, ::Val{:quote})

For indexed part select.

VerilogWriter.wireexpr — Method

wireexpr(expr, ::Val{:ref})

Bit select or slice of wires, e.g. x[1] and x[p:1].

VerilogWriter.wireexpr — Method

wireexpr(expr::Expr, ::Val{:&})

Unary & parses differently from |(wire) and ^(wire). What is &(wire) originally used for in Julia?

VerilogWriter.wireexpr — Method

wireexpr(expr::Expr)

Parse one wire expression. Be sure to put 'two e's' in wir'ee'xpr, not 'wirexpr'.

Syntax

Part of what can be done in Verilog can be accepted, such as din[7:0], (w1 + w2) << 5

<wirename>

One wire, without slicing or bit-selecting.

<val::Int>, <val::hex>, <val::bin>

Literals, e.g. 5, 0x1f, 0b10.

<wire> <op> <wire>, <op> <wire>

Unary and binary operators. For reduction operators (unary &, |, ^), since these are not operators in Julia, write it in the form of function call explicitly, i.e. &(wire), |(wire) instead of doing ^wire. Note that we use ^ as xor just as in Verilog, though this is not a xor operator in Julia.

<wire>[<wire>:<wire>], <wire>[<wire>]

Bit select and slicing as in Verilog/SystemVerilog. Does not reject even if both wires in [<wire>:<wire>] are not of constant value.

wire[<wire>-:<wire>]

Indexed part select. Does not reject even if <wire> after -: is not of constant value. As using - operator and : (quote) in Julia syntax, no spaces between <wire>(here)-:(and here)<wire> are allowed, and <wire> after -: should in most cases be put inside parentheses.

Note that because : (quote) is used inside quote, you (for now) cannot embed objects here through Metaprogramming.

e.g. w = (@wireexpr w); oneblock(:(x[A-:($w)] <= y)) is not allowed.

In such cases use constructors instead.

julia> e = Wireexpr(ipselm, @wireexpr(x), @wireexpr(A), @wireexpr(w));

julia> vshow(oneblock(:($(e) <= y)));
x[A -: w] <= y;
type: Alassign

Examples

julia> w = wireexpr(:(w)); vshow(w);
w
type: Wireexpr

julia> w = wireexpr(:(w1 & &(w2) )); vshow(w);
(w1 & &(w2))
type: Wireexpr

julia> w = wireexpr(:(w[i:0])); vshow(w);
w[i:0]
type: Wireexpr

julia> w = @wireexpr w[(P*Q)-:(R+10)]; vshow(w);
w[(P * Q) -: (R + 10)]
type: Wireexpr

VerilogWriter.wireexpr — Method

wireexpr(expr::QuoteNode)

For indexed part select.

VerilogWriter.wireexpr — Method

wireexpr(expr::Wireexpr)

Insertion through metaprogramming.

Example

julia> w = @wireexpr x + y;

julia> e = :(a + |($(w) & z));

julia> ans = wireexpr(e); vshow(ans);
(a + |(((x + y) & z)))
type: Wireexpr

VerilogWriter.wireexpr — Method

wireexpr(expr, ::T, ::Val{:call}) where {T <: binopvals}

Parse binary operators.

VerilogWriter.wireexpr — Method

wireexpr(expr, ::T, ::Val{:call}) where {T <: arityambigVals}

Disambiguate symbols in arityambigVals between unary and binary operators.

VerilogWriter.wireexpr — Method

wireexpr(expr, ::T, ::Val{:call}) where {T <: unaopvals}

Parse unary operators.

VerilogWriter.wtypesym — Method

wtypesym(wt::Symbol)

Given either of symbols @wire, @reg, @logic, return wire, reg, logic (::Wiretype), respectively.

VerilogWriter.@decloneline — Macro

@decloneline(arg)

Macro version of decloneline.

VerilogWriter.@decls — Macro

@decls(arg)

Macro version of decls.

VerilogWriter.@ifcontent — Macro

@ifcontent(arg)

Macro call.

VerilogWriter.@localparams — Macro

@localparams(arg)

Macro version.

VerilogWriter.@oneblock — Macro

@oneblock(arg)

Parse arg (which is AST) using macro. Uses reference to prevent arg being evaluated.

VerilogWriter.@onelocalparam — Macro

@onelocalparam(arg)

Macro call.

VerilogWriter.@oneparam — Macro

@oneparam(arg)

Macro call.

VerilogWriter.@parameters — Macro

@parameters(arg)

Macro call.

VerilogWriter.@portoneline — Macro

@portoneline(arg)

Macro version of portoneline.

VerilogWriter.@ports — Macro

@ports(arg)

Macro version of ports.

VerilogWriter.@ralways — Macro

@ralways(arg)

ralways macro version.

Not supposed to be used in most cases, @always and always are often better.

VerilogWriter.atypealways — Method

atypealways(x::Ifcontent)

Infer type of always-block (into alwaysff or alwayscomb).

VerilogWriter.atypealways — Method

atypealways(x::Vector{T}) where {T}

Infer type of always-block from a vector of T.

VerilogWriter.punsolved — Constant

value to be returned when parameter value cannot be solved.

VerilogWriter.Wireop — Method

(cls::Wireop)(uno::Int, dos::Int)

Operators in wbinop are now callable.

Used in calculating parameters, for Wireexpr is used inside parameter objects.

VerilogWriter.Wireop — Method

(cls::Wireop)(uno::Int)

uminus, whose type is Wireop is now callable.

Used in calculating parameters, for Wireexpr is used inside parameter objects.

VerilogWriter.paramcalc — Method

paramcalc(w::Wireexpr, ans)

Evaluate paramter whose rhs is w under ans returned from paramsolve.

VerilogWriter.paramcalc — Method

paramcalc(w::Wireexpr)

Evaluate w under empty environment, for debugging.

VerilogWriter.paramsolve — Method

paramsolve(prm::Parameters, lprm::Localparams)

Using prm and lprm, calculate all parameters' rhs, which is a Wireexpr, and determine the constant value corresponds to each parameter.

VerilogWriter.paramsolvecore! — Method

paramsolvecore!(n::String, ans, alldict)

Prepare for the computation of the parameter named n.

VerilogWriter.paramsolvecore_inner! — Method

paramsolvecore_inner!(w, ans::Dict{String, Int}, alldict::Dict{String, Wireexpr}, visited::Dict{String, Bool})

Return the constant value of w::Wireexpr.

If the value for w has not been calculated, recursively (with depth-first-search) evaluate the parameters which appear at rhs of w, and update ans.

VerilogWriter.fsmconv — Method

fsmconv(::Type{Case}, x::FSM, atype::Atype)

Accept as an argument atype to be used in assignments inside a case statement. Defaults to ff in the method fsmconv(::Type{Case}, x::FSM).

VerilogWriter.fsmconv — Method

fsmconv(::Type{Case}, x::FSM)

Convert FSM object into Case object.

VerilogWriter.fsmconv — Method

fsmconv(::Type{Ifcontent}, x::FSM)

Convert FSM logic to Ifcontent object.

When vshow(fsmconv(Ifcontent, x)) is evaluated one case block will be the only output.

VerilogWriter.fsmconv — Method

fsmconv(::Type{Localparams}, x::FSM)

Generate localparams that declare the value which corresponds to each state.

VerilogWriter.fsmconv — Method

fsmconv(::Type{Onedecl}, x::FSM)

Generate one reg declaration in verilog that holds the state value at the time.

VerilogWriter.transadd! — Method

transadd!(x::FSM, cond::Wireexpr, newtrans::Pair{String, String})

Add a new transition rule for the state machine x.

The new rule here is:

The transition from state newtrans[1] to state newtrans[2]

This transition occures when cond is true and the current state is newtrans[1]

Examples

julia> fsm = @FSM nstate (uno, dos, tres); # create a FSM

julia> transadd!(fsm, (@wireexpr b1 == b2), @tstate uno => dos); # transition from "uno" to "dos" when "b1 == b2"

julia> vshow(fsmconv(Case, fsm));
case (nstate)
    uno: begin
        if ((b1 == b2)) begin
            nstate <= dos;
        end
    end
    dos: begin
        
    end
    tres: begin
        
    end
endcase
type: Case

VerilogWriter.transcond — Method

transcond(m::FSM, states::Pair{String, String})

Get the wireexpr whose value should be true iff the state of the FSM changes from states[1] to states[2].

Examples

julia> m = @FSM fsm s1, s2, s3; transadd!(m, (@wireexpr b1 == TCOND), @tstate s1=>s2); vshow(fsmconv(Case, m));
case (fsm)
    s1: begin
        if ((b1 == TCOND)) begin
            fsm <= s2;
        end
    end
    s2: begin
        
    end
    s3: begin
        
    end
endcase
type: Case

julia> t = transcond(m, @tstate s1 => s2); vshow(t);
((fsm == s1) && (b1 == TCOND))
type: Wireexpr

julia> t = transcond(m, @tstate s2 => s1);
ERROR: transition rule not registered for "s2" => "s1".

VerilogWriter.@FSM — Macro

@FSM(mname, states)

Create FSM with variables of (machine itself / states) names.

With the macro there is no need to explicitly make a string object for machine/state names.

Example

julia> m = @FSM m1 (s1, s2); # same as "@FSM(m1, (s1, s2))"

julia> vshow(m)
reg m1;

localparam s1 = 0;
localparam s2 = 1;

case (m1)
    s1: begin
        
    end
    s2: begin
        
    end
endcase
type: FSM

VerilogWriter.@tstate — Macro

@tstate(arg)

Helper macro for argument in transadd!.

Convert a pair of variables to a pair of strings.

julia> @tstate a => b
"a" => "b"

julia> m = @FSM machine s1, s2;

julia> transadd!(m, (@wireexpr b), @tstate s1 => s2);

julia> vshow(m)
reg machine;

localparam s1 = 0;
localparam s2 = 1;

case (machine)
    s1: begin
        if (b) begin
            machine <= s2;
        end
    end
    s2: begin
        
    end
endcase
type: FSM

VerilogWriter.defclk — Constant

Default value of clocking signal.

VerilogWriter.defrst — Constant

Default value of resetting signal.

VerilogWriter.autoreset — Method

autoreset(x::Alwayscontent; clk=defclk, rst=defrst, edge=posedge)

Automatically reset wires which appear in x::Alwayscontent.

Sensitivity list in the original Alwayscontent will be ignored.

VerilogWriter.autoreset — Method

autoreset(x::Ifcontent; clk=Wireexpr("CLK"), rst=Wireexpr("RST"), edge=posedge)

Given x::Ifcontent, returns always_ff/always block that resets every wire/regs appear at Lhs of x.

Example

c = @ifcontent (
    r1 <= r2;
    if b1 
        r2 <= 0
        r3 <= r3 + r4
    else 
        r3 <= 0
    end
) 
r = autoreset(c; clk=(@wireexpr clk), rst=(@wireexpr ~resetn))
vshow(r)

# output

always_ff @( posedge clk ) begin
    if (~resetn) begin
        r1 <= 0;
        r2 <= 0;
        r3 <= 0;
    end else begin
        r1 <= r2;
        if (b1) begin
            r2 <= 0;
            r3 <= (r3 + r4);
        end else begin
            r3 <= 0;
        end
    end
end
type: Alwayscontent

VerilogWriter.lhsextract — Method

lhsextract(x::Alassign)

VerilogWriter.lhsextract — Method

lhsextract(x::Case)

VerilogWriter.lhsextract — Method

lhsextract(x::Ifcontent)

Extract all Wireexprs on Lhs. May contain more than one identical Wireexprs.

VerilogWriter.lhsextract — Method

lhsextract(x::Ifelseblock)

VerilogWriter.lhsextractcore! — Method

lhsextractcore!(ans, v)

A helper function for lhsextract.

VerilogWriter.lhsunify — Method

lhsunify(wvec::Vector{Wireexpr})

Remove identical wires contained in wvec more than once.

VerilogWriter.wirewidid — Constant

update julia to v1.8 and replace with wirewidid::Int = 0.

VerilogWriter.wwinvalid — Constant

Value which indicates that the Wirewid object does not have a valid value.

VerilogWriter.Vmodenv — Method

Vmodenv()

Create an empty Vmodenv object.

VerilogWriter.WirewidthUnresolved — Type

Error thrown when wire width inferenece is not possible.

VerilogWriter.autodecl — Method

autodecl(x, env::Vmodenv)

Declare wires in x which are not yet declared in env. Raise error when not enough information to determine width of all wires is given.

Type of x is restricted to the type wireextract! accepts.

Examples

Inference Success

pts = @ports (
        @in 16 din;
        @in b1
)
env = Vmodenv(pts)

c = @ifcontent (
    reg1 = 0;
    reg2 = din;
    if b1 
        reg1 = din[10:7]
    end
) 

newds, _ = autodecl(c, env)
vshow(newds)

# output

reg [3:0] reg1;
reg [15:0] reg2;
type: Decls

You may also declare ports/wires beforehand whose width is unknown.

When declaring ports/wires without specifying its bit width, assign -1 as its width.

ps = @ports (
    @in 2 x;
    @in -1 y;
    @out @reg A z
) 
ds = @decls (
    @wire -1 w1;
    @wire B w2
)

ab = @always (
    z <= r1 + r2 + r3;
    r4 <= (y & w1) << r1[1:0];
    r5 <= y + w2
)
env = Vmodenv(Parameters(), ps, Localparams(), ds)
d, newenv = autodecl(ab.content, env)

vshow(newenv)
println()
vshow(d)

# output

input [1:0] x
input [B-1:0] y
output reg [A-1:0] z

wire [B-1:0] w1;
wire [B-1:0] w2;
type: Vmodenv

reg [A-1:0] r1;
reg [A-1:0] r2;
reg [A-1:0] r3;
reg [B-1:0] r4;
reg [B-1:0] r5;
type: Decls

Fail in Inference

c = @ifcontent (
    reg1 = 0;
    reg2 = din;
    if b1 
        reg1 = din[10:7]
    end
) 

autodecl(c)

# output

ERROR: Wire width cannot be inferred for the following wires.
1. b1
2. reg2 = din

VerilogWriter.autodecl — Method

autodecl(x)

Call autodecl under an empty environment.

VerilogWriter.eqwidflatten! — Method

eqwidflatten!(x::Wireexpr, envdicts, ansset, eqwids::Vector{Wirewid}, declonly, equality)

Extract wires that appear inside x which are of the same width as x itself, find Wirewid objects which correspond to the wire and push them into eqwids. When slice appears inside x, new Wirewid object is created and then added to eqwids.

VerilogWriter.extract2dreg — Method

extract2dreg(x::Vector{Onedecl})

Extract 2d regs as dict object from x.

VerilogWriter.findandpush_widunify! — Method

findandpush_widunify!(newval::String, envdicts, ansset, widvars)

Looking inside envdicts, add newval to ansset and widvars if newval is not in envdicts.

VerilogWriter.portdeclupdated! — Method

portdeclupdated!(env::Vmodenv, ansset::Dict{String, Wirewid})

When env contains ports/wire declarations of unknown width, return new Vmodenv object whose bitwidth are all filled in.

ansset is supposed to contain enough information.

VerilogWriter.separate_widunknown — Method

separate_widunknown(v::Vector{Onedecl})

Separate Onedecls into the group of wires whose width is already determined and the group of wires with unknown width.

VerilogWriter.strwidunknown — Method

strwidunknown(widvars)

Format widvars, returned from widunify, to a string.

VerilogWriter.unifycore_widunify! — Method

unifycore_widunify!(items::Vector{T}, envdicts, ansset, widvars, declonly, equality) where {T <: Union{Wireexpr, Tuple{Wireexpr, Wireexpr}}}

Called in widunify, infer wire width from equality constraints.

To recursively look into Wireexprs in items, push new equality and declaration into declonly and equality given as arguments.

VerilogWriter.unknowndeclpush! — Method

unknowndeclpush!(ansset::Dict{String, Wirewid}, widvars::Dict{Wirewid, Vector{String}}, ukports::Ports, ukdecls::Decls)

Register ports/wire declarations of unknown width to ansset and widvars.

VerilogWriter.widunify — Method

widunify(declonly::Vector{Wireexpr}, equality::Vector{Tuple{Wireexpr, Wireexpr}}, env::Vmodenv)

Given declonly and equality from wireextract, infer width of wires which appear in the conditions.

VerilogWriter.widunify — Method

widunify(declonly::Vector{Wireexpr}, equality::Vector{Tuple{Wireexpr, Wireexpr}})

Call widunify under an empty env (, where no ports, parameters,... are declared).

VerilogWriter.wireextract! — Method

wireextract!(x::Alwayscontent, declonly, equality)

For always content. Sensitivity in x will be ignored.

VerilogWriter.wireextract! — Method

wireextract!(x::Ifcontent, declonly, equality)

Helper function for wireextract. Add constraints to declonly and equality.

VerilogWriter.wireextract! — Method

wireextract!(x::Vector{T}, declonly, equality) where {T <: Union{Ifcontent, Alwayscontent}}

For Vector{Ifcontent/Alwayscontent}.

VerilogWriter.wireextract — Method

wireextract(x)

Extract Wireexprs from x::Ifcontent, along with equality constraints of wire width.

Return declonly::Vector{Wireexpr}, which contains wires appear in x::Ifcontent, and equality::Vector{Tuple{Wireexpr, Wireexpr}}, which contains tuples of two Wireexprs which are supposed to be of the same width.

Type of x is restricted in wireextract!.

VerilogWriter.wirepush_widunify! — Method

wirepush_widunify!(w::Wireexpr, envdicts, ansset, widvars)

Push every wire that appears inside w to ansset and widvars.