Component Structs

Structs Description

This module offers several structs that contain a structure of Verilog components, major ones of which are listed below.

Parameters

Oneparam

Ports

Oneport

Portdirec

Wiretype

Localparams

Onelocalparam

Wireexpr

Wireop

Decls

Onedecl

Assign

Alwayscontent

Alassign

Ifelseblock

Ifcontent

Atype

Edge

Vmodinst

Vmodule

Parameters

Parameters(val::Vector{Oneparam})

Missing docstring.

Missing docstring for Parameters. Check Documenter's build log for details.

Oneparam

Oneparam(name::String, val::Wireexpr)

Missing docstring.

Missing docstring for Oneparam. Check Documenter's build log for details.

Ports

Ports(val::Vector{Oneport})

Missing docstring.

Missing docstring for Ports. Check Documenter's build log for details.

Oneport

Oneport(direc::Portdirec, decl::Onedecl)

Missing docstring.

Missing docstring for Oneport. Check Documenter's build log for details.

Portdirec

Enum pin pout

Missing docstring.

Missing docstring for Portdirec. Check Documenter's build log for details.

Wiretype

Enum wire reg logic

Missing docstring.

Missing docstring for Wiretype. Check Documenter's build log for details.

Localparams

Localparams(val::Vector{Onelocalparam})

Missing docstring.

Missing docstring for Localparams. Check Documenter's build log for details.

Onelocalparam

Onelocalparam(name::String, val::Wireexpr)

Missing docstring.

Missing docstring for Onelocalparam. Check Documenter's build log for details.

Wireexpr

Wireexpr(operation::Wireop, name::String, subnodes::Vector{Wireexpr}, bitwidth::Int64, value::Int64)

Missing docstring.

Missing docstring for Wireexpr. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Wireexpr(n::String). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Wireexpr(n::Symbol). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Wireexpr(n::Int). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Wireexpr(op::Wireop, w::Wireexpr...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Wireexpr(op::Wireop, v::Vector{Wireexpr}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Wireexpr(n::String, msb::T) where {T <: Union{Int, Wireexpr}}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Wireexpr(n::String, msb::T1, lsb::T2) where {T1 <: Union{Int, Wireexpr}, T2 <: Union{Int, Wireexpr}}. Check Documenter's build log for details.

VerilogWriter.Wireexpr — Method

Wireexpr(w::Integer, n::Integer)

Literal with width specification, printed in the decimal format.

Missing docstring.

Missing docstring for Wireexpr(expr::Wireexpr). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Wireexpr(). Check Documenter's build log for details.

Wireop

Enum add minus mul vdiv lshift rshift band bor bxor neg uminus redand redor redxor logieq leq lt land lor id slice literal ipselm

Missing docstring.

Missing docstring for Wireop. Check Documenter's build log for details.

Decls

Decls(val::Vector{Onedecl})

Missing docstring.

Missing docstring for Decls. Check Documenter's build log for details.

Onedecl

Onedecl(wtype::Wiretype, width::Wireexpr, name::String, is2d::Bool, wid2d::Wireexpr)

Missing docstring.

Missing docstring for Onedecl. Check Documenter's build log for details.

Assign

Assign(lhs::Wireexpr, rhs::Wireexpr)

Missing docstring.

Missing docstring for Assign. Check Documenter's build log for details.

Alwayscontent

Alwayscontent(atype::Atype, sens::Sensitivity, content::Ifcontent)

Missing docstring.

Missing docstring for Alwayscontent. Check Documenter's build log for details.

Alassign

Alassign(lhs::Wireexpr, rhs::Wireexpr, atype::Atype)

Missing docstring.

Missing docstring for Alassign. Check Documenter's build log for details.

Ifelseblock

Ifelseblock(conds::Vector{Wireexpr}, contents::Vector{Ifcontent})

VerilogWriter.Ifelseblock — Type

Container of an if-else block. Parametrized by T for mutual recursion. Used as Ifelseblock = Ifelseblock_inner{Case}.

Ifcontent

Ifcontent(assigns::Vector{Alassign}, ifelseblocks::Vector{Ifelseblock}, cases::Vector{Case})

VerilogWriter.Ifcontent — Type

Container of one if-block (, one elseif block, or one else block).

Parametrized by T, U only for mutual recursion with Ifelseblock and Case, thus used as Ifcontent_inner{Ifelseblock, Case}, which is aliased as Ifcontent.

Atype

Enum ff comb aunknown

Missing docstring.

Missing docstring for Atype. Check Documenter's build log for details.

Edge

Enum posedge negedge unknownedge

Missing docstring.

Missing docstring for Edge. Check Documenter's build log for details.

Vmodinst

Vmodinst(vmodname::String, instname::String, params::Vector{Pair{String, Wireexpr}}, ports::Vector{Pair{String, Wireexpr}}, wildconn::Bool)

Missing docstring.

Missing docstring for Vmodinst. Check Documenter's build log for details.

Vmodule

Vmodule(name::String, params::Parameters, ports::Ports, locparams::Localparams, decls::Decls, insts::Vector{Vmodinst}, assigns::Vector{Assign}, always::Vector{Alwayscontent})

Missing docstring.

Missing docstring for Vmodule. Check Documenter's build log for details.

Converter Functions/Macros

As in previous examples we offer functions and macros to convert Julia syntax into certain structs described above. You may use these instead of calling constructors.

All the functions listed below accept Expr object as its argument (e.g. :(x = 10), see Julia Documents for more information.), and that is why variables inside the argument Expr object do not have to be declared anywhere else in the source code. The syntaxes each function requires in a argument Expr objects are also described below (or may be easily inferred from the examples here and Brief Introduction).

As you see in Brief Introduction there are macros that do the same thing as functions listed below (and both macro and function has the same name). As macros take Expr object as its argument, you can write codes in a slightly more simple manner with macros. For example,

always(:(
    @posedge clk; 

    w1 <= w2;
    if b1 == b2 
        w3 <= w4
    end
))

is equivalent to

@always (
    @posedge clk; 

    w1 <= w2;
    if b1 == b2 
        w3 <= w4
    end
)

(be care full not to foreget (one space) between macros and (, i.e. @macro(a,b) and @macro (a,b) are different.)

But sometimes there are things what macros cannot do (for now), an example is having for loop inside expressions.

d = always(:(
    if b1 
        $([:($(Symbol("x$i")) = $(Symbol("y$i"))) for i in 1:3]...)
    end
))
vshow(d)

and this outputs

always_comb begin
    if (b1) begin
        x1 = y1;
        x2 = y2;
        x3 = y3;
    end
end
type: Alwayscontent

List of Converter Functions/Macros

Written inside parentheses are the types of objects the functions return.

parameters (Parameters)

oneparam (Oneparam)

ports (Ports)

portoneline (Oneport)

wireexpr (Wireexpr)

localparams (Localparams)

onelocalparam (Onelocalparam)

decls (Decls)

decloneline (Onedecl)

always (Alwayscontent)

oneblock (Ifelseblock,Alassign)

ifcontent

parameters

Missing docstring.

Missing docstring for parameters(expr::Expr). Check Documenter's build log for details.

oneparam

Missing docstring.

Missing docstring for oneparam(expr::Expr). Check Documenter's build log for details.

ports

Missing docstring.

Missing docstring for ports(::Expr). Check Documenter's build log for details.

Missing docstring.

Missing docstring for ports(::Vector{Oneport}). Check Documenter's build log for details.

portoneline

Missing docstring.

Missing docstring for portoneline(::Expr). Check Documenter's build log for details.

wireexpr

Missing docstring.

Missing docstring for wireexpr(::Expr). Check Documenter's build log for details.

Missing docstring.

Missing docstring for wireexpr(::Wireexpr). Check Documenter's build log for details.

localparams

Missing docstring.

Missing docstring for localparams(expr::Expr). Check Documenter's build log for details.

onelocalparam

Missing docstring.

Missing docstring for onelocalparam(expr::Expr). Check Documenter's build log for details.

decls

Missing docstring.

Missing docstring for decls(::Expr). Check Documenter's build log for details.

Missing docstring.

Missing docstring for decls(::Vector{Onedecl}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for decls(expr::Decls...). Check Documenter's build log for details.

decloneline

Missing docstring.

Missing docstring for decloneline(::Expr). Check Documenter's build log for details.

always

Missing docstring.

Missing docstring for always(::Expr). Check Documenter's build log for details.

oneblock

Missing docstring.

Missing docstring for oneblock(::Expr). Check Documenter's build log for details.

Missing docstring.

Missing docstring for oneblock(expr::T) where {T <: Union{Alassign, Ifelseblock}}. Check Documenter's build log for details.

ifcontent

Missing docstring.

Missing docstring for ifcontent(x::Expr). Check Documenter's build log for details.

Embedding Objects

You can embed generated objects back into Verilog-like codes. Note that because we ask you to make use of metaprogramming (interpolation in particular), macros cannot be used for the purpose.

By embedding objects as Julia AST, you can construct new objects that contain the information of embedded objects.

Every object (offered in this package) can be embedded almost anywhere it seems to be possible.

Examples

julia> a = @portoneline @in clk;

julia> b = ports(:($(a); @out 8 dout)); vshow(b);
(
    input clk,
    output [7:0] dout
);
type: Ports

julia> a = @ports (@in clk; @out 8 dout);

julia> b = ports(:(@in resetn; $(a))); vshow(b);
(
    input resetn,
    input clk,
    output [7:0] dout
);
type: Ports

julia> a = @wireexpr (x + y) & z;

julia> b = always(:(lhs = $(a) | w)); vshow(b);
always_comb begin
    lhs = (((x + y) & z) | w);
end
type: Alwayscontent

Other Functions

vshow

Missing docstring.

Missing docstring for vshow(::Any). Check Documenter's build log for details.

invports

VerilogWriter.invports — Function

invports(ps::Ports)

Return Ports object whose directions are reversed from ps.

Examples

pts = @ports (
    @in 8 bus1, bus2;
    @out @reg bus3
)
ipts = invports(pts)
vshow(pts)
vshow(ipts)

# output

(
    input [7:0] bus1,
    input [7:0] bus2,
    output reg bus3
);
type: Ports
(
    output [7:0] bus1,
    output [7:0] bus2,
    input bus3
);
type: Ports

declmerge

VerilogWriter.declmerge — Function

declmerge(d::Parameters...)

Merge multiple Parameters objects into one Parameters.

declmerge(d::Decls...)

Merge multiple Decls objects into one Decls.

declmerge(d::Localparams...)

Merge multiple Localparams objects into one Localparams.

declmerge(d::Ports...)

Merge multiple Ports objects into one Ports.

sym2wire

VerilogWriter.@sym2wire — Macro

@sym2wire(arg::Symbol)

Declare new Wireexpr of name arg.

@sym2wire(arg::Expr)

Declare new Wireexprs of name arg..., respectively.

Example

julia> @sym2wire x, y, z;

julia> vshow(y);
y
type: Wireexpr