Automatic Inference

We offer some tools to automatically add additional information inferred from a given Verilog-like codes.

Reset in Always Blocks

Given a content of always blocks, you may automatically reset all wires which appear at the LHS in the block.

VerilogWriter.Core.autoreset — Function

autoreset(x::Ifcontent; clk=defclk, rst=defrst, edge=posedge, reg2d::Dict{String, Wireexpr}=Dict{String, Wireexpr}())

Given x::Ifcontent, returns always_ff/always block that resets every wire/regs appear at Lhs of x.

This is synchronous reset. reg2d is a pair of "name of wire" and "ram depth in wireexpr".

autoreset(x::Alwayscontent; clk=defclk, rst=defrst, edge=posedge, reg2d::Dict{String, Wireexpr}=Dict{String, Wireexpr}())

Automatically reset wires which appear in x::Alwayscontent.

Sensitivity list in the original Alwayscontent will be ignored.

Example

julia> c = @always (
            r1 <= r2;
            if b1 
                r2 <= 0
                r3 <= r3 + r4
            else 
                r3 <= 0
            end
       );

julia> r = autoreset(c; clk=(@wireexpr clk), rst=(@wireexpr ~resetn));

julia> vshow(r);
always_ff @( posedge clk ) begin
    if ((~resetn)) begin
        r1 <= 0;
        r2 <= 0;
        r3 <= 0;
    end else begin
        r1 <= r2;
        if (b1) begin
            r2 <= 0;
            r3 <= (r3 + r4);
        end else begin
            r3 <= 0;
        end
    end
end
type: Alwayscontent

autoreset(x::Vmodule; clk=defclk, rst=defrst, edge=posedge)

Return a new Vmodule object whose Alwayscontents are all reset.

In some cases reg/logic should not be reset (e.g. 2d array is initialized with $readmemh and must not initialized in always block). To generate Alwayscontent that should not be reset by autoreset you may call @nralways.

VerilogWriter.Core.@nralways — Macro

@nralways(arg)

Generate Alwayscontent whose noreset field is set to true.

julia> alReset = @always (
            mem[inindex] <= din;
       );

julia> autoreset(alReset) |> vshow;
always_ff @( posedge CLK ) begin
    if (RST) begin
        mem <= 0;
    end else begin
        mem[inindex] <= din;
    end
end
type: Alwayscontent

julia> alNoreset = @nralways (
            mem[inindex] <= din;
       );

julia> autoreset(alNoreset) |> vshow;
always_ff @( posedge CLK ) begin
    mem[inindex] <= din;
end
type: Alwayscontent

Automatic Wire Declaration

VerilogWriter.Core.autodecl — Function

autodecl(x, env::Vmodenv, instInterfaces::VinstInterfaces)::Tuple{WidthInferenceStatus, Vmodenv}

Declare wires in x which are not yet declared in env. Raise error when not enough information to determine width of all wires is given.

Examples

Inference Success

julia> pts = @ports (
       @in 16 din;
       @in b1
       );

julia> env = Vmodenv(pts);

julia> c = @ifcontent (
       reg1 = 0;
       reg2 = din;
       if b1 
       reg1 = din[10:7]
       end
       ); 

julia> status, venv = autodecl(c, env); vshow(venv, true);
input b1
input [15:0] din

logic [3:0] reg1;
logic [15:0] reg2;
type: Vmodenv

julia> widthInferenceCompleted(status)
true

You may also declare ports/wires beforehand whose width is unknown.

When declaring ports/wires without specifying its bit width, assign -1 as its width.

julia> ps = @ports (
       @in 2 x;
       @in -1 y;
       @out @reg A z
       );

julia> ds = @decls (
       @wire -1 w1;
       @wire B w2
       );

julia> ab = @always (
       z <= r1 + r2 + r3;
       r4 <= (y & w1) << r1[1:0];
       r5 <= y + w2
       );

julia> env = Vmodenv(Parameters(), ps, Localparams(), ds);

julia> status, nenv = autodecl(ab.content, env);

julia> vshow(nenv, true);
input [1:0] x
input [B-1:0] y
output reg [A-1:0] z

logic [A-1:0] r1;
logic [A-1:0] r2;
logic [A-1:0] r3;
logic [B-1:0] r4;
logic [B-1:0] r5;
wire [B-1:0] w1;
wire [B-1:0] w2;
type: Vmodenv

julia> widthInferenceCompleted(status)
true

Fail in Inference

julia> c = @always (
       reg1 = 0;
       reg2 = din;
       if b1 
           reg1 = din[10:7]
       end
       );

julia> status, _ = autodecl(c); throwIfInferenceIsIncomplete(status);
ERROR: Wire width cannot be inferred for the following wires.
1. b1
2. din = reg2

autodecl(x)

Conduct wire width inference under an empty environment.

autodecl(x::Vmodule)::Tuple{WidthInferenceStatus, Vmodule}

Using ports, parameters, localparams, decls in x::Vmodule as an environment, conduct wire width inference and return a new Vmodule object with inferred wires.

VerilogWriter.Core.autodeclVmodlist — Function

autodeclVmodlist(x::Vector{Vmodule}, getFixedPoint::Bool)

Apply autodecl to each Vmodule object in x. Port width inferred in one Vmodule object (say A) is considered when inferring wire width of another Vmodule object in which A is instantiated.

When getFixedPoint is false, port width which can actually be inferred from nested module instantiation may not be fully inferred.

autodeclVmodlist(vmods)

When not specified otherwise getFixedPoint is set to true.

VerilogWriter.Core.WidthInferenceStatus — Type

Status to be returned after width inference.

VerilogWriter.Core.widthInferenceCompleted — Function

widthInferenceCompleted(status::WidthInferenceStatus)

Return true if width inference is fully done, false otherwise (i.e. width of some wires / ports remain unresolved).

VerilogWriter.Core.throwIfInferenceIsIncomplete — Function

throwIfInferenceIsIncomplete(status::WidthInferenceStatus)

Throw some error according to status if width inference is not complete.

VerilogWriter.Core.vfinalize — Function

vfinalize(x::Vmodule)

Return new Vmodule object generated by applying autoreset and autodecl on x.

vfinalize(x::Vector{Vmodule})

Return new Vmodule objects generated by applying autoreset and autodeclVmodlist on x.