Patch

We offer some methods to generate group of wires / always blocks with certain functionality.

These methods return Vpatch object which contains wire declarations and always blocks so that wires returned together with the Vpatch object can achieve an expected behavior.

VerilogWriter.Core.Vpatch — Type

Wrapper object to add logics and wire declarations to Vmodule.

Example of using patch methods:

julia> resetStdpatchCounter(); # Restart suffix of generated wire from one

julia> w, p = isAtRisingEdge(@wireexpr(data)); vshow(w); # returns a wire and a `Vpatch` object. `w` is a wire which is set to 1 iff `data` is at rising edge.
_risingEdge_1
type: Wireexpr

julia> v = Vmodule("test"); vpush!(v, p); vshow(v); # `Vpatch` object can be pushed into `Vmodule` object.
module test ();
    always_comb begin
        _risingEdge_1 = ((_prev_risingEdge_1 == 1'd0) & (data == 1'd1));
    end
    always_ff @( unknownedge  ) begin
        _prev_risingEdge_1 <= data;
    end
endmodule
type: Vmodule

Wires generated in patch methods are named with care to avoid generating wires with identical name. To achieve the goal most of these methods take name::AbstractString as the last argument.

Usually an Integer value which is incremented every time a method is called is passed to the name argument. To make methods declaration simple, we offer an helper macro @vstdpatch.

VerilogWriter.Core.@vstdpatch — Macro

@vstdpatch(expr)

Attached at methods generating a pair of something (usually wires) and Vpatch. Methods should take name::AbstractString as the last argument. The macro additionally defines the method which does not take name argument, and which internally passes to name an Int value which is incremented everytime the method without name argument is called.

To summarize, when this macro takes

function Foo(arg1::T1, arg2::T2, name::AbstractString)
    ...
end

as its argument, it generates

# The original method definition
function Foo(arg1::T1, arg2::T2, name::AbstractString)
    ...
end

# Additional method without `name`
function Foo(arg1::T1, arg2::T2)
    return Foo(arg1, arg2, string(incrementStdpatchCounter()))
end

Basic Patches

VerilogWriter.Core.posedgePrec — Function

posedgePrec(earlier::Wireexpr, later::Wireexpr, name::AbstractString)

Return wireexpr which indicates if a rising edge in earlier occured earlier than that of later.

The wires earlier and later being high from the beginning is regarded as an edge at the beginning.

Return Wire

(1): 1 if edge detected in either of two wires

(0): 1 if edge in earlier was earlier than later

VerilogWriter.Core.bitbundle — Function

bitbundle(wvec::Vector{Wireexpr}, name::AbstractString)

Return wire which bundles Wireexprs in wvec.

Width of wires in wvec are supposed to be all one.

VerilogWriter.Core.nonegedge — Function

nonegedge(uno::Wireexpr, name::AbstractString)

Return wire which shows whether wire uno underwent a falling edge.

Wire uno must be single-bit wide or fails in width inference.

Return Wire

(0): 1 if wire uno has never encountered a falling edge

VerilogWriter.Core.posedgeSync — Function

posedgeSync(uno::Wireexpr, dos::Wireexpr, name::AbstractString)

Return Wireexpr which indicates whether a rising edge is detected at the same clock cycle in uno and dos.

Return Wire

(1): 1 if edge detected in either of two wires

(0): 1 if edge in earlier was earlier than later

VerilogWriter.Core.invertBitOrder — Function

invertBitOrder(w::Wireexpr, wid::Int)

Invert bit order of wire w.

Return Wire

(wid-1:0): Bit order of w is inverted.

VerilogWriter.Core.isAtRisingEdge — Function

isAtRisingEdge(w::Wireexpr, name::AbstractString)

Return wire whose value is 1 iff the wire w is at the rising edge.

Return Wire

(0): 1 at the cycle at which a rising edge is detected.

VerilogWriter.Core.interceptBuffer — Function

interceptBuffer(data::Wireexpr, update::Wireexpr, name::AbstractString)

Buffer that changes its data at the exact cycle at which update is triggerred.

Should be careful on the critical path on using this buffer.

Return Wire

(length(data)-1:0): Buffer output

VerilogWriter.Core.onceHigh — Function

onceHigh(wire::Wireexpr, restart::Wireexpr)

Check if wire is once asserted.

Input Wires

restart

(0): Restart checking if wire is asserted.

Return wire

(0): 1 if wire is asserted once after restart is deasserted.

VerilogWriter.Core.zipSpike — Function

zipSpike(wires::Vector{Wireexpr})

Zip rising edges of wires in wires.

Return Wire

(0): 1 if all wire in wires experienced rising edge after this wire is previously asserted.

RAM Related Patches

VerilogWriter.Core.readOnlyQueueComb — Function

readOnlyQueueComb(width::Int, data::V, update, restart, name::AbstractString) where {T,V<:AbstractVector{T}}

Generate queue which outputs value in data in the order which they are aligned in data. Implemented with 1d reg and always_comb.

Each item in data is converted to a wire in verilog with width width.

Input Wires

update : Wire to be asserted to request next data to the queue.

restart : When asserted the queue starts outputting value in data from the first item.

Return Wires

dataValid

(0): High indicates that the output from the queue is valid.

outData

(width-1:0): Data from the queue.

VerilogWriter.Core.readOnlyQueue — Function

readOnlyQueue(width::Int, depth::Int, update, restart, memfilename, name::AbstractString)

Generate a buffer queue whose data is initalized with memfile using initial statement.

Output is same as that of readOnlyQueueComb.

VerilogWriter.Core.fifoPatch — Function

fifoPatch(depth, width, inputData, inputValid, updateOutput, name::AbstractString)

Implement FIFO in a so-called first-word-fall-through style.

Input Wires

inputData

(width-1:0): Data input to FIFO.

inputValid

(0): 1 if inputData contains valid data.

updateOutput

(0): 1 when requesting next data to FIFO. After updateOutput is set to 1, dout would contain valid data at the first cycle at which outValid is set to 1.

Return Wires

inReady

(0): 1 if fifo is ready to accept new input.

outValid

(0): 1 if dout contains valid data.

dout

(width-1:0): Data output from FIFO.