Simulation with Verilator

We offer a simple method to run simulation on verilog modules using verilator (verilator must be installed beforehand).

VerilogWriter.Core.verilatorSimrun — Function

verilatorSimrun(resultbuf::IO, dut::V, tplen::Integer, cycles::Integer; option::VerilatorOption=DEFAULT_VERILATOR_OPTION) where {V <: Union{Vmodule, Vector{Vmodule}}}

Run simulation using Verilator.

dut must pass acceptableInVerilatorTest. Simulation is run for cycles cycles, for tplen-many test points. Active-high reset signal RST is currently set to 0 20 clock cycles after starting the simulation.

Bits in the port named tp in dut is supposed to be all high at the end of simulation. true is returned if tp == '1, false otherwise.

Files that contain test codes for a simulation with verilator is generated in a temporary directory tmpverilatorsim, which is automatically generated in the current directory when the simulation is run.

verilatorSimrun(dut, tplen, cycles; option::VerilatorOption=DEFAULT_VERILATOR_OPTION)

Run verilatorSimrun with the result being output to STDOUT.

VerilogWriter.Core.acceptableInVerilatorTest — Function

acceptableInVerilatorTest(v::Vmodule, tplen::Integer)::Bool

Check that ports in v are one bit wide CLK and RST, and tplen wide tp. Additionally, name of v should be dut.

VerilatorOption(wall::Bool, disabledWarning::Vector{String})

VerilogWriter.Core.VerilatorOption — Type

Represents options passed to verilator.