Finite State Machine

We offer here an easy way to construct a Finite State Machine (FSM).

Finite State Machine is a common structure used in a hardware design. Offering a simple method to construct a FSM may be useful.

Procedure of Designing a FSM

Construct a FSM

First construct a FSM defining name of states and the machine itself.

VerilogWriter.@FSM — Macro

@FSM(mname, states)

Create FSM with variables of (machine itself / states) names.

With the macro there is no need to explicitly make a string object for machine/state names.

Example

julia> m = @FSM m1 (s1, s2); # same as "@FSM(m1, (s1, s2))"

julia> vshow(m)
reg m1;

localparam s1 = 0;
localparam s2 = 1;

case (m1)
    s1: begin
        
    end
    s2: begin
        
    end
endcase
type: FSM

VerilogWriter.FSM — Method

FSM(name, state::String...)

Create an FSM object with name name, whose states are state....

VerilogWriter.FSM — Method

FSM(name, states::Vector{String})

Create an FSM object with name name, whose states are states.

Add Transition Rules

Add a new rule with transadd!.

VerilogWriter.transadd! — Function

transadd!(x::FSM, cond::Wireexpr, newtrans::Pair{String, String})

Add a new transition rule for the state machine x.

The new rule here is:

The transition from state newtrans[1] to state newtrans[2]

This transition occures when cond is true and the current state is newtrans[1]

Examples

julia> fsm = @FSM nstate (uno, dos, tres); # create a FSM

julia> transadd!(fsm, (@wireexpr b1 == b2), @tstate uno => dos); # transition from "uno" to "dos" when "b1 == b2"

julia> vshow(fsmconv(Case, fsm));
case (nstate)
    uno: begin
        if ((b1 == b2)) begin
            nstate <= dos;
        end
    end
    dos: begin
        
    end
    tres: begin
        
    end
endcase
type: Case

transadd!(x::FSM, rules::Vector{Tuple{Wireexpr, Pair{String, String}}})

Add muliple rules to x.

As shown above cond is the condition to be true when the transition occurs, and as newtrans argument you assign a Pair object of strings, "srcstate" => "deststate". You may instead call macro @tstate srcstate => deststate.

VerilogWriter.@tstate — Macro

@tstate(arg)

Helper macro for argument in transadd!.

Convert a pair of variables to a pair of strings.

julia> @tstate a => b
"a" => "b"

julia> m = @FSM machine s1, s2;

julia> transadd!(m, (@wireexpr b), @tstate s1 => s2);

julia> vshow(m)
reg machine;

localparam s1 = 0;
localparam s2 = 1;

case (machine)
    s1: begin
        if (b) begin
            machine <= s2;
        end
    end
    s2: begin
        
    end
endcase
type: FSM

Generate Condition to be True when Transition Occurs

After adding transition rules you may generate wireexpr which would be true only when the transition occurs.

VerilogWriter.transcond — Function

transcond(m::FSM, states::Pair{String, String})

Get the wireexpr whose value should be true iff the state of the FSM changes from states[1] to states[2].

Examples

julia> m = @FSM fsm s1, s2, s3; transadd!(m, (@wireexpr b1 == TCOND), @tstate s1=>s2); vshow(fsmconv(Case, m));
case (fsm)
    s1: begin
        if ((b1 == TCOND)) begin
            fsm <= s2;
        end
    end
    s2: begin
        
    end
    s3: begin
        
    end
endcase
type: Case

julia> t = transcond(m, @tstate s1 => s2); vshow(t);
((fsm == s1) && (b1 == TCOND))
type: Wireexpr

julia> t = transcond(m, @tstate s2 => s1);
ERROR: transition rule not registered for "s2" => "s1".

Convert from FSM to Verilog Codes

We offer methods to generate several components in verilog, which compose FSM structure.

They are

A case statement for transition between states

reg declaration that holds the current state information.

localparams each of which represents one state in the FSM.

VerilogWriter.fsmconv — Function

fsmconv(::Type{Case}, x::FSM)

Convert FSM object into Case object.

fsmconv(::Type{Case}, x::FSM, atype::Atype)

Accept as an argument atype to be used in assignments inside a case statement. Defaults to ff in the method fsmconv(::Type{Case}, x::FSM).

fsmconv(::Type{Ifcontent}, x::FSM)

Convert FSM logic to Ifcontent object.

When vshow(fsmconv(Ifcontent, x)) is evaluated one case block will be the only output.

fsmconv(::Type{Onedecl}, x::FSM)

Generate one reg declaration in verilog that holds the state value at the time.

fsmconv(::Type{Decls}, x::FSM)

Generate Decls object from x.

fsmconv(::Type{Localparams}, x::FSM)

Generate localparams that declare the value which corresponds to each state.

vpush! is available to push all information above to a Vmodule object. You may also prewiew all of these with vshow(x::FSM).

Struct FSM Description

FSM(name::String, states::Vector{String}, transcond::Matrix{Wireexpr}, transvalid::Matrix{Bool})

VerilogWriter.FSM — Type

Struct for a Finite State Machine, not the Flying Spaghetti Monster.

Each column of matrices corresponds to the source state, and row for the destination state. Not supposed to directly edit the content of matrices without using handlers.

[
    false false 
    true  false
]

in transvalid means transition from the 1st state to the 2nd state is provided.