Basic Automation

We offer some tools to automatically add additional information inferred from a given Verilog-like codes.

Reset in Always Blocks

Given a content of always blocks, you may automatically reset all wires which appear at the LHS in the block.

VerilogWriter.autoreset — Function

autoreset(x::Ifcontent; clk=Wireexpr("CLK"), rst=Wireexpr("RST"), edge=posedge)

Given x::Ifcontent, returns always_ff/always block that resets every wire/regs appear at Lhs of x.

Example

c = @ifcontent (
    r1 <= r2;
    if b1 
        r2 <= 0
        r3 <= r3 + r4
    else 
        r3 <= 0
    end
) 
r = autoreset(c; clk=(@wireexpr clk), rst=(@wireexpr ~resetn))
vshow(r)

# output

always_ff @( posedge clk ) begin
    if (~resetn) begin
        r1 <= 0;
        r2 <= 0;
        r3 <= 0;
    end else begin
        r1 <= r2;
        if (b1) begin
            r2 <= 0;
            r3 <= (r3 + r4);
        end else begin
            r3 <= 0;
        end
    end
end
type: Alwayscontent

Automatic Wire Declaration

VerilogWriter.autodecl — Function

autodecl(x::Ifcontent, env::Vmodenv)

Declare wires in x::Ifcontent which are not yet declared in env. Raise error when not enough information to determine width of all wires is given.

Examples

Inference Success

pts = @ports (
        @in 16 din;
        @in b1
)
env = Vmodenv(Parameters(), pts, Localparams(), Decls())

c = @ifcontent (
    reg1 = 0;
    reg2 = din;
    if b1 
        reg1 = din[10:7]
    end
) 

newds = autodecl(c, env)
vshow(newds)

# output

reg [3:0] reg1;
reg [15:0] reg2;
type: Decls

Fail in Inference

c = @ifcontent (
    reg1 = 0;
    reg2 = din;
    if b1 
        reg1 = din[10:7]
    end
) 

autodecl(c)

# output

ERROR: Wire width cannot be inferred for the following wires.
1. b1
2. reg2 = din

autodecl(x::Ifcontent)

Call autodecl under an empty environment.

env in an argument for autodecl is of type Vmodenv.

Vmodenv(prms::Parameters, prts::Ports, lprms::Localparams, dcls::Decls)

VerilogWriter.Vmodenv — Type

Environment in which wire width inference is done.