Component Structs

Structs Description

This module offers several structs that contain a structure of Verilog components, major ones of which are listed below.

Parameters

Oneparam

Ports

Oneport

Portdirec

Wiretype

Localparams

Onelocalparam

Wireexpr

Wireop

Decls

Onedecl

Assign

Alwayscontent

Alassign

Ifelseblock

Ifcontent

Atype

Edge

Vmodinst

Vmodule

Parameters

Parameters(val::Vector{Oneparam})

VerilogWriter.Parameters — Type

List of parameters.

Oneparam

Oneparam(name::String, val::Wireexpr)

VerilogWriter.Oneparam — Type

Parameter in verilog.

Ports

Ports(val::Vector{Oneport})

VerilogWriter.Ports — Type

Gather multiple ports.

Oneport

Oneport(direc::Portdirec, decl::Onedecl)

VerilogWriter.Oneport — Type

Represent a single port declaration.

Portdirec

Enum pin pout

VerilogWriter.Portdirec — Type

Port direction object.

Wiretype

Enum wire reg logic

VerilogWriter.Wiretype — Type

Wiretype object.

Localparams

Localparams(val::Vector{Onelocalparam})

VerilogWriter.Localparams — Type

Multiple localparams.

Onelocalparam

Onelocalparam(name::String, val::Wireexpr)

VerilogWriter.Onelocalparam — Type

One localparam.

Wireexpr

Wireexpr(operation::Wireop, name::String, subnodes::Vector{Wireexpr}, bitwidth::Int64, value::Int64)

VerilogWriter.Wireexpr — Type

Wire expressions in verilog.

Contains unnecessary information to handle all the wires in the same type.

One motivation to avoid making different types of wires (e.g. using parametric types) is that it seemed beneficial to have multiple wires in a vector, and that it is sometime a disadvantage in performance to use abstract types when creating vectors.

Some operators on wires in Verilog, which are listed in Enum Wireop, are overloaded for Wireexpr. Note that reduction operators (unary &, |, ^) are not in Julia, and logical and, or (&&, ||) can be applied only for booleans in Julia, thus are not available as an operator for Wireexpr objects.

Examples

julia> w1 = @wireexpr x + y;

julia> w2 = @wireexpr z;

julia> vshow(w1 & w2);
((x + y) & z)
type: Wireexpr

julia> w3 = @wireexpr x;

julia> w4 = @wireexpr y && z; # && is available inside `@wireexpr` and `wireexpr` methods.

julia> vshow(w3 | w4)
(x | (y && z))
type: Wireexpr

VerilogWriter.Wireexpr — Method

Wireexpr(n::String)

Case where no slice is needed.

VerilogWriter.Wireexpr — Method

Wireexpr(n::Symbol)

Convert Symbol to String and construct Wireexpr.

VerilogWriter.Wireexpr — Method

Wireexpr(n::Int)

Literal of Integer.

VerilogWriter.Wireexpr — Method

Wireexpr(op::Wireop, w::Wireexpr...)

Apply an operation of w wires.

Examples

julia> a = @wireexpr a; b = @wireexpr b;

julia> c = Wireexpr(add, a, b); # equivalent to wireexpr(:($(a) + $(b)))

julia> vshow(c);
(a + b)
type: Wireexpr

julia> d = Wireexpr(redor, a); vshow(d); # wireexpr(:(|($(a))))
|(a)
type: Wireexpr

VerilogWriter.Wireexpr — Method

Wireexpr(op::Wireop, v::Vector{Wireexpr})

Apply an operation on wires in v.

VerilogWriter.Wireexpr — Method

Wireexpr(n::String, msb::T) where {T <: Union{Int, Wireexpr}}

Get one bit at msb from n.

VerilogWriter.Wireexpr — Method

Wireexpr(n::String, msb::T1, lsb::T2) where {T1 <: Union{Int, Wireexpr}, T2 <: Union{Int, Wireexpr}}

Slice of n from msb to lsb.

Index for wires can be Int or Wireexpr.

VerilogWriter.Wireexpr — Method

Wireexpr(w::Int, n::Int)

Literal with width specification, printed in the decimal format.

VerilogWriter.Wireexpr — Method

Wireexpr(expr::Wireexpr)

Return the argument itself, used in slice construction to make it possible to apply the same method to msb as Int and as Wireexpr.

VerilogWriter.Wireexpr — Method

Wireexpr()

Create an empty expression.

Wireop

Enum add minus mul vdiv lshift rshift band bor bxor neg uminus redand redor redxor logieq leq lt land lor id slice literal ipselm

VerilogWriter.Wireop — Type

Verilog operators.

Unary &, | does not exist so explicitly call as function e.g. &(wire), |(wire) (& behaves in a wickedmanner...?), and are only available inside quoted expressions. Xor in verilog ^ is in Julia exponential operator, and the difference in an association exists.

Decls

Decls(val::Vector{Onedecl})

VerilogWriter.Decls — Type

Multiple wire declarations.

Onedecl

Onedecl(wtype::Wiretype, width::Wireexpr, name::String)

VerilogWriter.Onedecl — Type

Represent one wire declaration.

Assign

Assign(lhs::Wireexpr, rhs::Wireexpr)

VerilogWriter.Assign — Type

Assign one statement.

Alwayscontent

Alwayscontent(atype::Atype, sens::Sensitivity, content::Ifcontent)

VerilogWriter.Alwayscontent — Type

Represent always blocks.

Alassign

Alassign(lhs::Wireexpr, rhs::Wireexpr, atype::Atype)

VerilogWriter.Alassign — Type

Assign statement inside always blocks.

Ifelseblock

Ifelseblock(conds::Vector{Wireexpr}, contents::Vector{Ifcontent})

VerilogWriter.Ifelseblock — Type

Container of an if-else block. Parametrized by T for mutual recursion. Used as Ifelseblock = Ifelseblock_inner{Case}.

Ifcontent

Ifcontent(assigns::Vector{Alassign}, ifelseblocks::Vector{Ifelseblock}, cases::Vector{Case})

VerilogWriter.Ifcontent — Type

Container of one if-block (, one elseif block, or one else block).

Parametrized by T, U only for mutual recursion with Ifelseblock and Case, thus used as Ifcontent_inner{Ifelseblock, Case}, which is aliased as Ifcontent.

Atype

Enum ff comb aunknown

VerilogWriter.Atype — Type

Type of always blocks.

Edge

Enum posedge negedge unknownedge

VerilogWriter.Edge — Type

Edge in sensitivity lists.

Vmodinst

Vmodinst(vmodname::String, instname::String, params::Vector{Pair{String, Wireexpr}}, ports::Vector{Pair{String, Wireexpr}}, wildconn::Bool)

VerilogWriter.Vmodinst — Type

Vmodinst

Represents module instantiation.

i = Vmodinst(
    "mod1",
    "u1",
    [
        "A" => Wireexpr(15)
    ],
    [
        "p1" => (@wireexpr x << 1),
        "p2" => (@wireexpr y)
    ]
)
vshow(i)

# output

mod1 #(
    .A(15)
) u1 (
    .p1((x << 1)),
    .p2(y)
);
type: Vmodinst

Vmodule

Vmodule(name::String, params::Parameters, ports::Ports, locparams::Localparams, decls::Decls, insts::Vector{Vmodinst}, assigns::Vector{Assign}, always::Vector{Alwayscontent})

VerilogWriter.Vmodule — Type

Represents one verilog module.

Converter Functions/Macros

As in previous examples we offer functions and macros to convert Julia syntax into certain structs described above. You may use these instead of calling constructors.

All the functions listed below accept Expr object as its argument (e.g. :(x = 10), see Julia Documents for more information.), and that is why variables inside the argument Expr object do not have to be declared anywhere else in the source code. The syntaxes each function requires in a argument Expr objects are also described below (or may be easily inferred from the examples here and Brief Introduction).

As you see in Brief Introduction there are macros that do the same thing as functions listed below (and both macro and function has the same name). As macros take Expr object as its argument, you can write codes in a slightly more simple manner with macros. For example,

always(:(
    @posedge clk; 

    w1 <= w2;
    if b1 == b2 
        w3 <= w4
    end
))

is equivalent to

@always (
    @posedge clk; 

    w1 <= w2;
    if b1 == b2 
        w3 <= w4
    end
)

(be care full not to foreget (one space) between macros and (, i.e. @macro(a,b) and @macro (a,b) are different.)

But sometimes there are things what macros cannot do (for now), an example is having for loop inside expressions.

d = always(:(
    if b1 
        $([:($(Symbol("x$i")) = $(Symbol("y$i"))) for i in 1:3]...)
    end
))
vshow(d)

and this outputs

always_comb begin
    if (b1) begin
        x1 = y1;
        x2 = y2;
        x3 = y3;
    end
end
type: Alwayscontent

List of Converter Functions/Macros

Written inside parentheses are the types of objects the functions return.

parameters (Parameters)

oneparam (Oneparam)

ports (Ports)

portoneline (Oneport)

wireexpr (Wireexpr)

localparams (Localparams)

onelocalparam (Onelocalparam)

decls (Decls)

decloneline (Onedecl)

always (Alwayscontent)

oneblock (Ifelseblock,Alassign)

ifcontent

parameters

VerilogWriter.parameters — Method

parameters(expr::Expr)

Multiple parameters.

Using convert and localparams inside.

Example

v = parameters(:(
    x = 10;
    y = 20;
    z = 30
))
vshow(v)

# output

#(
    parameter x = 10,
    parameter y = 20,
    parameter z = 30
)
type: Parameters

oneparam

VerilogWriter.oneparam — Method

oneparam(expr::Expr)

One parameter.

Syntax

<paramname> = <val::Int>

julia> vshow(oneparam(:(x = 10)));
parameter x = 10
type: Oneparam

ports

VerilogWriter.ports — Method

ports(expr::Expr)

Convert Julia AST into port declarations as Ports object.

Syntax

<portoneline>[;<portoneline>;...]

Multiple lines of portoneline expressions separated by ; can be accepted.

Example

pp = ports(:(
    @in p1;
    @in @wire 8 p2, p3, p4;
    @out @reg 2 p5, p6
))

vshow(pp)

# output

(
    input p1,
    input [7:0] p2,
    input [7:0] p3,
    input [7:0] p4,
    output reg [1:0] p5,
    output reg [1:0] p6
);
type: Ports

VerilogWriter.ports — Method

ports(expr::Vector{Oneport})

Insertion through metaprogramming.

Example

a = @portoneline @in 8 d1, d2, d3
b = ports(:(
    @in d0;
    $(a);
    @out @reg 8 dout
))
vshow(b)

# output

(
    input d0,
    input [7:0] d1,
    input [7:0] d2,
    input [7:0] d3,
    output reg [7:0] dout
);
type: Ports

portoneline

VerilogWriter.portoneline — Method

portoneline(expr::Expr)

Parse Julia AST to one line of port declaration as Oneport.

Syntax

@in <wirename>, @out <wirename>

One port declaration of width 1.

@in/@out <wirename1, wirename2, ...>

Multiple port declaration of width 1 in one line.

@in/@out <width> <wirename1[, wirename2, ...]>

Port declarations of width <width>.

@in/@out @<wiretype> [<width>] <wirename1[, wirename2, ...]>

Port declaration with wiretypes [of width <width>].

Examples

julia> p1 = portoneline(:(@in din));

julia> vshow(p1);
input din
type: Oneport

julia> p2 = portoneline(:(@out din1, din2, din3)); vshow(p2);
output din1
type: Oneport
output din2
type: Oneport
output din3
type: Oneport

julia> p3 = portoneline(:(@in 8 din)); vshow(p3);
input [7:0] din
type: Oneport

julia> p4 = portoneline(:(@out @reg 8 dout)); vshow(p4);
output reg [7:0] dout
type: Oneport

julia> p5 = portoneline(:(@out (A+B)<<2 x, y)); vshow(p5); # width with parameter
output [((A + B) << 2)-1:0] x
type: Oneport
output [((A + B) << 2)-1:0] y
type: Oneport

wireexpr

VerilogWriter.wireexpr — Method

wireexpr(expr::Expr)

Parse one wire expression. Be sure to put 'two e's' in wir'ee'xpr, not 'wirexpr'.

Syntax

Part of what can be done in Verilog can be accepted, such as din[7:0], (w1 + w2) << 5

<wirename>

One wire, without slicing or bit-selecting.

<val::Int>, <val::hex>, <val::bin>

Literals, e.g. 5, 0x1f, 0b10.

<wire> <op> <wire>, <op> <wire>

Unary and binary operators. For reduction operators (unary &, |, ^), since these are not operators in Julia, write it in the form of function call explicitly, i.e. &(wire), |(wire) instead of doing ^wire. Note that we use ^ as xor just as in Verilog, though this is not a xor operator in Julia.

<wire>[<wire>:<wire>], <wire>[<wire>]

Bit select and slicing as in Verilog/SystemVerilog. Does not reject even if both wires in [<wire>:<wire>] are not of constant value.

wire[<wire>-:<wire>]

Indexed part select. Does not reject even if <wire> after -: is not of constant value. As using - operator and : (quote) in Julia syntax, no spaces between <wire>(here)-:(and here)<wire> are allowed, and <wire> after -: should in most cases be put inside parentheses.

Note that because : (quote) is used inside quote, you (for now) cannot embed objects here through Metaprogramming.

e.g. w = (@wireexpr w); oneblock(:(x[A-:($w)] <= y)) is not allowed.

In such cases use constructors instead.

julia> e = Wireexpr(ipselm, @wireexpr(x), @wireexpr(A), @wireexpr(w));

julia> vshow(oneblock(:($(e) <= y)));
x[A -: w] <= y;
type: Alassign

Examples

julia> w = wireexpr(:(w)); vshow(w);
w
type: Wireexpr

julia> w = wireexpr(:(w1 & &(w2) )); vshow(w);
(w1 & &(w2))
type: Wireexpr

julia> w = wireexpr(:(w[i:0])); vshow(w);
w[i:0]
type: Wireexpr

julia> w = @wireexpr w[(P*Q)-:(R+10)]; vshow(w);
w[(P * Q) -: (R + 10)]
type: Wireexpr

VerilogWriter.wireexpr — Method

wireexpr(expr::Wireexpr)

Insertion through metaprogramming.

Example

julia> w = @wireexpr x + y;

julia> e = :(a + |($(w) & z));

julia> ans = wireexpr(e); vshow(ans);
(a + |(((x + y) & z)))
type: Wireexpr

localparams

VerilogWriter.localparams — Method

localparams(expr::Expr)

Multiple localparams.

Examples

julia> p = localparams(:(x = 10)); vshow(p);
localparam x = 10;
type: Localparams

p = localparams(:(
    a = 111;
    b = 222;
    c = 333
))
vshow(p)

# output

localparam a = 111;
localparam b = 222;
localparam c = 333;
type: Localparams

onelocalparam

VerilogWriter.onelocalparam — Method

onelocalparam(expr::Expr)

One localparam object. The syntax is the same as oneparam.

Example

julia> vshow(onelocalparam(:(x = 100)))
localparam x = 100;
type: Onelocalparam

decls

VerilogWriter.decls — Method

decls(expr::Expr)

Parse Julia AST into wire declaration section object Decls.

Syntax

<onedecl>[;<onedecl>;...]

Multiple decloneline expressions which are concatenated by ; can be accepted.

Example

d = decls(:(
    @wire w1;
    @reg 8 w2,w3,w4;
    @logic 32 w5
))
vshow(d)

# output

wire w1;
reg [7:0] w2;
reg [7:0] w3;
reg [7:0] w4;
logic [31:0] w5;
type: Decls

VerilogWriter.decls — Method

decls(expr::Vector{Onedecl})

For insertion with metaprogramming.

Example

a = @decloneline @reg 8 x1, x2
b = decls(:(
    $(a);
    @wire y1, y2
))
vshow(b)

# output

reg [7:0] x1;
reg [7:0] x2;
wire y1;
wire y2;
type: Decls

VerilogWriter.decls — Method

decls(expr::Decls...)

For interpolation of Decls objects.

decloneline

VerilogWriter.decloneline — Method

decloneline(expr::Expr)

Parse Julia AST into wire declaration as Vector{Onedecl}.

The number of Onedecl objects returned may differ according to the number of wires declared in one line (e.g. input dout <=> input din1, din2, din3).

Syntax

Similar to that of portoneline.

@wire/@reg/@logic [<width>] <wirename1>[, <wirename2>,...]

Examples

julia> d = decloneline(:(@reg 10 d1)); vshow(d);
reg [9:0] d1;
type: Onedecl

julia> d = decloneline(:(@logic 8 d1,d2,d3)); vshow(d);
logic [7:0] d1;
type: Onedecl
logic [7:0] d2;
type: Onedecl
logic [7:0] d3;
type: Onedecl

julia> d = decloneline(:(@wire A >> 2 w1, w2)); vshow(d);
wire [(A >> 2)-1:0] w1;
type: Onedecl
wire [(A >> 2)-1:0] w2;
type: Onedecl

always

VerilogWriter.always — Method

always(expr::Expr)

Parse AST into always block as Alwayscontent using ralways.

Also infers type of always using addatype!.

Syntax

<oneblock>;<oneblock>[;<oneblock>;...]

<oneblock> is the expression that can be parsed by oneblock. ; in between <oneblock>s are strictly needed.

@posedge <wirename>; <ifelsestatements>/<assignments>

Set sensitivity list using macro syntax. @negedge is also possible. You must put @posegde/@negedge statement at the beginning, and only once.

Examples

a1 = always(:(
    w1 = w2;
    if b2 
        w1 = w3 
    end
))
vshow(a1)

# output

always_comb begin
    w1 = w2;
    if (b2) begin
        w1 = w3;
    end
end
type: Alwayscontent

a1 = always(:(
    @posedge clk;
    
    if b1 == b2
        w1 <= w2 + w3 
    else
        w1 <= ~w1 
    end
))
vshow(a1)

# output

always_ff @( posedge clk ) begin
    if ((b1 == b2)) begin
        w1 <= (w2 + w3);
    end else begin
        w1 <= ~w1;
    end
end
type: Alwayscontent

oneblock

VerilogWriter.oneblock — Method

oneblock(expr::Expr)

Parse AST in Julia grammar to one if-else statement or one assignment inside always blocks in Verilog, which are Ifelseblock and Alassign, respectively.

As using Julia expression and Julia parser as if it is of Verilog, there can be difference in grammatical matters between what can be given as the expr and real Verilog (e.g. operator precedence of <=).

Syntax

<wirename1> = <wireoperation>

One blocking assignment. (Note that bit width cannot be specified yet, nor can it be printed.) <wireoperation> is a expression accepted by wireexpr.

If-else statement

if <wireoperation>
    <oneassignment>
    <oneassignment>
    ...
elseif <wireoperation> 
    <oneassignment>
    <ifelsestatement>
    ...
else
    <ifelsestatement>
    ...
end

If-else statement written in 'Julia syntax', not in Verilog syntax, can be accepted. else block and elseif are not compulsory. Since if end are at the top level no ; inside if-else statement is needed. Nested if-else statement can be also accepted as in usual Julia.

Examples

julia> a1 = oneblock(:(w1 <= w2)); 

julia> vshow(a1);
w1 <= w2;
type: Alassign

julia> a2 = oneblock(:(w3 = w4 + ~w2)); vshow(a2);
w3 = (w4 + ~w2);
type: Alassign

a3 = oneblock(:(
    if b1 == b2
        w5 = ~w6
        w7 = w8 
    elseif b2 
        w9 = w9 + w10
    else
        if b3 
            w11 = w12 
        end
    end
))
vshow(a3)

# output

if ((b1 == b2)) begin
    w5 = ~w6;
    w7 = w8;
end else if (b2) begin
    w9 = (w9 + w10);
end else begin
    if (b3) begin
        w11 = w12;
    end
end
type: Ifelseblock

VerilogWriter.oneblock — Method

oneblock(expr::T) where {T <: Union{Alassign, Ifelseblock}}

For insertion through metaprogramming.

Example

a = @oneblock r = s & t
b = @oneblock (
    if b 
        x = y 
    else
        x = z
    end
)
c = always(:(
    p = q;
    $(a);
    $(b)
))
vshow(c)

# output

always_comb begin
    p = q;
    r = (s & t);
    if (b) begin
        x = y;
    end else begin
        x = z;
    end
end
type: Alwayscontent

ifcontent

VerilogWriter.ifcontent — Method

ifcontent(x::Expr)

Convert into Ifcontent what is convertible to Alwayscontent.

Example

x = @ifcontent (
    a = b;
    if b 
        x = c
    else
        x = d 
    end
)
vshow(x)

# output

a = b;
if (b) begin
    x = c;
end else begin
    x = d;
end
type: Ifcontent

Embedding Objects

You can embed generated objects back into Verilog-like codes. Note that because we ask you to make use of metaprogramming (interpolation in particular), macros cannot be used for the purpose.

By embedding objects as Julia AST, you can construct new objects that contain the information of embedded objects.

Every object (offered in this package) can be embedded almost anywhere it seems to be possible.

Examples

julia> a = @portoneline @in clk;

julia> b = ports(:($(a); @out 8 dout)); vshow(b);
(
    input clk,
    output [7:0] dout
);
type: Ports

julia> a = @ports (@in clk; @out 8 dout);

julia> b = ports(:(@in resetn; $(a))); vshow(b);
(
    input resetn,
    input clk,
    output [7:0] dout
);
type: Ports

julia> a = @wireexpr (x + y) & z;

julia> b = always(:(lhs = $(a) | w)); vshow(b);
always_comb begin
    lhs = (((x + y) & z) | w);
end
type: Alwayscontent

Other Functions

vshow

VerilogWriter.vshow — Method

vshow(x; systemverilog=true)

Print the structs in VerilogWriter.jl in a readable format (except for enums). When systemverilog=false, output always, always @* instead of always_ff, always_comb, respectively.

invports

VerilogWriter.invports — Function

invports(ps::Ports)

Return Ports object whose directions are reversed from ps.

Examples

pts = @ports (
    @in 8 bus1, bus2;
    @out @reg bus3
)
ipts = invports(pts)
vshow(pts)
vshow(ipts)

# output

(
    input [7:0] bus1,
    input [7:0] bus2,
    output reg bus3
);
type: Ports
(
    output [7:0] bus1,
    output [7:0] bus2,
    input bus3
);
type: Ports

declmerge

VerilogWriter.declmerge — Function

declmerge(d::Parameters...)

Merge multiple Parameters objects into one Parameters.

declmerge(d::Decls...)

Merge multiple Decls objects into one Decls.

declmerge(d::Localparams...)

Merge multiple Localparams objects into one Localparams.

declmerge(d::Ports...)

Merge multiple Ports objects into one Ports.